Duplicated code and constants removal
