+++ src/bolos/cx_const_m25519.c
@@ -1,0 +1,68 @@
+
+/* ------------------------------------------------------------------------ */
+/* ---                             Curve25519                           --- */
+/* ------------------------------------------------------------------------ */
+
+static uint8_t const C_cx_Curve25519_a[] = {
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x07, 0x6d, 0x06
+};
+
+static uint8_t const C_cx_Curve25519_b[] = {
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01
+};
+static uint8_t const C_cx_Curve25519_p[] = {
+  // p:  0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffed
+  0x7f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xed
+};
+static uint8_t const C_cx_Curve25519_Hp[] = {
+  // Hp: 0x00000000000000000000000000000000000000000000000000000000000005a4
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0xa4
+};
+static uint8_t const C_cx_Curve25519_Gx[] = {
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x09
+};
+static uint8_t const C_cx_Curve25519_Gy[] = {
+  0x5f, 0x51, 0xe6, 0x5e, 0x47, 0x5f, 0x79, 0x4b, 0x1f, 0xe1, 0x22,
+  0xd3, 0x88, 0xb7, 0x2e, 0xb3, 0x6d, 0xc2, 0xb2, 0x81, 0x92, 0x83,
+  0x9e, 0x4d, 0xd6, 0x16, 0x3a, 0x5d, 0x81, 0x31, 0x2c, 0x14
+};
+static uint8_t const C_cx_Curve25519_n[] = {
+  // n:  0x1000000000000000000000000000000014DEF9DEA2F79CD65812631A5CF5D3ED
+  0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0xDE, 0xF9, 0xDE, 0xA2, 0xF7,
+  0x9C, 0xD6, 0x58, 0x12, 0x63, 0x1A, 0x5C, 0xF5, 0xD3, 0xED
+};
+static uint8_t const C_cx_Curve25519_Hn[] = {
+  // Hn: 0x0399411b7c309a3dceec73d217f5be65d00e1ba768859347a40611e3449c0f01
+  0x03, 0x99, 0x41, 0x1b, 0x7c, 0x30, 0x9a, 0x3d, 0xce, 0xec, 0x73,
+  0xd2, 0x17, 0xf5, 0xbe, 0x65, 0xd0, 0x0e, 0x1b, 0xa7, 0x68, 0x85,
+  0x93, 0x47, 0xa4, 0x06, 0x11, 0xe3, 0x44, 0x9c, 0x0f, 0x01
+};
+
+#define C_cx_Curve25519_h C_cofactor_8
+
+cx_curve_montgomery_t const C_cx_Curve25519 = {
+  .curve = CX_CURVE_Curve25519,
+  .bit_size = 256,
+  .length = 32,
+
+  .a = C_cx_Curve25519_a,
+  .b = C_cx_Curve25519_b,
+  .p = C_cx_Curve25519_p,
+  .Gx = C_cx_Curve25519_Gx,
+  .Gy = C_cx_Curve25519_Gy,
+  .n = C_cx_Curve25519_n,
+  .h = C_cx_Curve25519_h,
+  .Hp = C_cx_Curve25519_Hp,
+  .Hn = C_cx_Curve25519_Hn,
+};
