+++ src/bolos/cx_const_m448.c
@@ -1,0 +1,91 @@
+
+
+/* ------------------------------------------------------------------------ */
+/* ---                              Curve448                            --- */
+/* ------------------------------------------------------------------------ */
+
+static uint8_t const C_cx_Curve448_a[] = {
+  // a: 0x262a6
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x62, 0xa6
+};
+static uint8_t const C_cx_Curve448_b[] = {
+  // b:  0x01
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01
+};
+static uint8_t const C_cx_Curve448_p[] = {
+  // p:
+  // 0xfffffffffffffffffffffffffffffffffffffffffffffffffffffffeffffffffffffffffffffffffffffffffffffffffffffffffffffffff
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+  0xff, 0xff, 0xff, 0xfe, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff
+};
+static uint8_t const C_cx_Curve448_Hp[] = {
+  // Hp:
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+};
+static uint8_t const C_cx_Curve448_Gx[] = {
+  // Gx:
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05
+};
+static uint8_t const C_cx_Curve448_Gy[] = {
+  // Gy:
+  // 0x7d235d1295f5b1f66c98ab6e58326fcecbae5d34f55545d060f75dc28df3f6edb8027e2346430d211312c4b150677af76fd7223d457b5b1a
+  0x7d, 0x23, 0x5d, 0x12, 0x95, 0xf5, 0xb1, 0xf6, 0x6c, 0x98, 0xab, 0x6e,
+  0x58, 0x32, 0x6f, 0xce, 0xcb, 0xae, 0x5d, 0x34, 0xf5, 0x55, 0x45, 0xd0,
+  0x60, 0xf7, 0x5d, 0xc2, 0x8d, 0xf3, 0xf6, 0xed, 0xb8, 0x02, 0x7e, 0x23,
+  0x46, 0x43, 0x0d, 0x21, 0x13, 0x12, 0xc4, 0xb1, 0x50, 0x67, 0x7a, 0xf7,
+  0x6f, 0xd7, 0x22, 0x3d, 0x45, 0x7b, 0x5b, 0x1a
+};
+static uint8_t const C_cx_Curve448_n[] = {
+  // n:
+  // 0x3fffffffffffffffffffffffffffffffffffffffffffffffffffffff7cca23e9c44edb49aed63690216cc2728dc58f552378c292ab5844f3
+  0x3f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+  0xff, 0xff, 0xff, 0xff, 0x7c, 0xca, 0x23, 0xe9, 0xc4, 0x4e, 0xdb, 0x49,
+  0xae, 0xd6, 0x36, 0x90, 0x21, 0x6c, 0xc2, 0x72, 0x8d, 0xc5, 0x8f, 0x55,
+  0x23, 0x78, 0xc2, 0x92, 0xab, 0x58, 0x44, 0xf3
+};
+static uint8_t const C_cx_Curve448_Hn[] = {
+  // Hn:
+  0x1a, 0x9c, 0xc1, 0x4b, 0xa3, 0xc4, 0x7c, 0x44, 0xae, 0x17, 0xcf, 0x72,
+  0xc9, 0x86, 0x0f, 0xaa, 0x10, 0x75, 0x50, 0xba, 0x94, 0xd0, 0xe8, 0x08,
+  0xf1, 0x0a, 0x98, 0xef, 0x88, 0xe1, 0x5b, 0xb4, 0x8b, 0xd8, 0x27, 0xde,
+  0x12, 0x7c, 0xe5, 0xf2, 0x3e, 0x0a, 0xa6, 0x8d, 0x30, 0xdb, 0xb9, 0x89,
+  0xbb, 0xec, 0xff, 0xc3, 0x3c, 0x3a, 0x37, 0x8c
+};
+
+#define C_cx_Curve448_h C_cofactor_4
+
+cx_curve_montgomery_t const C_cx_Curve448 = {
+  .curve = CX_CURVE_Curve448,
+  .bit_size = 448,
+  .length = 56,
+
+  .a = C_cx_Curve448_a,
+  .b = C_cx_Curve448_b,
+  .p = C_cx_Curve448_p,
+  .Gx = C_cx_Curve448_Gx,
+  .Gy = C_cx_Curve448_Gy,
+  .n = C_cx_Curve448_n,
+  .h = C_cx_Curve448_h,
+  .Hp = C_cx_Curve448_Hp,
+  .Hn = C_cx_Curve448_Hn,
+};
