--- src/bolos/cx_ec_domain.c
@@ -1,1554 +1,0 @@
-#define _SDK_2_0_
-#include <err.h>
-#include <errno.h>
-#include <openssl/bn.h>
-#include <openssl/objects.h>
-#include <openssl/rand.h>
-#include <stdbool.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <time.h>
-#include <unistd.h>
-
-#include "bolos/cxlib.h"
-
-/* ========================================================================= */
-/* ========================================================================= */
-/* ===                                ALL                                === */
-/* ========================================================================= */
-/* ========================================================================= */
-
-static uint8_t const C_cofactor_1[] = { 0, 0, 0, 1 };
-static uint8_t const C_cofactor_4[] = { 0, 0, 0, 4 };
-static uint8_t const C_cofactor_8[] = { 0, 0, 0, 8 };
-
-/* ========================================================================= */
-/* ========================================================================= */
-/* ===                         SHORT WEIERSTRASS                         === */
-/* ========================================================================= */
-/* ========================================================================= */
-
-/* ------------------------------------------------------------------------ */
-/* ---                            secp256k1                             --- */
-/* ------------------------------------------------------------------------ */
-
-static uint8_t const C_cx_secp256k1_a[] = {
-  // a:  0x00
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-};
-static uint8_t const C_cx_secp256k1_b[] = {
-  // b:  0x07
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x07
-};
-static uint8_t const C_cx_secp256k1_p[] = {
-  // p:  0xfffffffffffffffffffffffffffffffffffffffffffffffffffffffefffffc2f
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0xff, 0xff, 0xfc, 0x2f
-};
-static uint8_t const C_cx_secp256k1_Hp[] = {
-  // Hp: 0x000000000000000000000000000000000000000000000001000007a2000e90a1
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x01, 0x00, 0x00, 0x07, 0xa2, 0x00, 0x0e, 0x90, 0xa1
-};
-static uint8_t const C_cx_secp256k1_Gx[] = {
-  // Gx: 0x79be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798
-  0x79, 0xbe, 0x66, 0x7e, 0xf9, 0xdc, 0xbb, 0xac, 0x55, 0xa0, 0x62,
-  0x95, 0xce, 0x87, 0x0b, 0x07, 0x02, 0x9b, 0xfc, 0xdb, 0x2d, 0xce,
-  0x28, 0xd9, 0x59, 0xf2, 0x81, 0x5b, 0x16, 0xf8, 0x17, 0x98
-};
-static uint8_t const C_cx_secp256k1_Gy[] = {
-  // Gy:  0x483ada7726a3c4655da4fbfc0e1108a8fd17b448a68554199c47d08ffb10d4b8
-  0x48, 0x3a, 0xda, 0x77, 0x26, 0xa3, 0xc4, 0x65, 0x5d, 0xa4, 0xfb,
-  0xfc, 0x0e, 0x11, 0x08, 0xa8, 0xfd, 0x17, 0xb4, 0x48, 0xa6, 0x85,
-  0x54, 0x19, 0x9c, 0x47, 0xd0, 0x8f, 0xfb, 0x10, 0xd4, 0xb8
-};
-static uint8_t const C_cx_secp256k1_n[] = {
-  // n: 0xfffffffffffffffffffffffffffffffebaaedce6af48a03bbfd25e8cd0364141
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xfe, 0xba, 0xae, 0xdc, 0xe6, 0xaf, 0x48,
-  0xa0, 0x3b, 0xbf, 0xd2, 0x5e, 0x8c, 0xd0, 0x36, 0x41, 0x41
-};
-static uint8_t const C_cx_secp256k1_Hn[] = {
-  // Hn:0x9d671cd581c69bc5e697f5e45bcd07c6741496c20e7cf878896cf21467d7d140
-  0x9d, 0x67, 0x1c, 0xd5, 0x81, 0xc6, 0x9b, 0xc5, 0xe6, 0x97, 0xf5,
-  0xe4, 0x5b, 0xcd, 0x07, 0xc6, 0x74, 0x14, 0x96, 0xc2, 0x0e, 0x7c,
-  0xf8, 0x78, 0x89, 0x6c, 0xf2, 0x14, 0x67, 0xd7, 0xd1, 0x40
-};
-
-#define C_cx_secp256k1_h C_cofactor_1
-
-cx_curve_weierstrass_t const C_cx_secp256k1 = {
-  .curve = CX_CURVE_SECP256K1,
-  .bit_size = 256,
-  .length = 32,
-  .a = C_cx_secp256k1_a,
-  .b = C_cx_secp256k1_b,
-  .p = C_cx_secp256k1_p,
-  .Gx = C_cx_secp256k1_Gx,
-  .Gy = C_cx_secp256k1_Gy,
-  .n = C_cx_secp256k1_n,
-  .h = C_cx_secp256k1_h,
-  .Hp = C_cx_secp256k1_Hp,
-  .Hn = C_cx_secp256k1_Hn,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                            secp256r1                             --- */
-/* ------------------------------------------------------------------------ */
-
-static uint8_t const C_cx_secp256r1_a[] = {
-  // a: 0xffffffff00000001000000000000000000000000fffffffffffffffffffffffc
-  0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfc
-};
-static uint8_t const C_cx_secp256r1_b[] = {
-  // b: 0x5ac635d8aa3a93e7b3ebbd55769886bc651d06b0cc53b0f63bce3c3e27d2604b
-  0x5a, 0xc6, 0x35, 0xd8, 0xaa, 0x3a, 0x93, 0xe7, 0xb3, 0xeb, 0xbd,
-  0x55, 0x76, 0x98, 0x86, 0xbc, 0x65, 0x1d, 0x06, 0xb0, 0xcc, 0x53,
-  0xb0, 0xf6, 0x3b, 0xce, 0x3c, 0x3e, 0x27, 0xd2, 0x60, 0x4b
-};
-static uint8_t const C_cx_secp256r1_p[] = {
-  // p:  0xffffffff00000001000000000000000000000000ffffffffffffffffffffffff
-  0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff
-};
-static uint8_t const C_cx_secp256r1_Hp[] = {
-  // Hp: 0x00000004fffffffdfffffffffffffffefffffffbffffffff0000000000000003
-  0x00, 0x00, 0x00, 0x04, 0xff, 0xff, 0xff, 0xfd, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xfe, 0xff, 0xff, 0xff, 0xfb, 0xff, 0xff,
-  0xff, 0xff, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03
-};
-static uint8_t const C_cx_secp256r1_Gx[] = {
-  // Gx: 0x6b17d1f2e12c4247f8bce6e563a440f277037d812deb33a0f4a13945d898c296
-  0x6b, 0x17, 0xd1, 0xf2, 0xe1, 0x2c, 0x42, 0x47, 0xf8, 0xbc, 0xe6,
-  0xe5, 0x63, 0xa4, 0x40, 0xf2, 0x77, 0x03, 0x7d, 0x81, 0x2d, 0xeb,
-  0x33, 0xa0, 0xf4, 0xa1, 0x39, 0x45, 0xd8, 0x98, 0xc2, 0x96
-};
-static uint8_t const C_cx_secp256r1_Gy[] = {
-  // Gy: 0x4fe342e2fe1a7f9b8ee7eb4a7c0f9e162bce33576b315ececbb6406837bf51f5
-  0x4f, 0xe3, 0x42, 0xe2, 0xfe, 0x1a, 0x7f, 0x9b, 0x8e, 0xe7, 0xeb,
-  0x4a, 0x7c, 0x0f, 0x9e, 0x16, 0x2b, 0xce, 0x33, 0x57, 0x6b, 0x31,
-  0x5e, 0xce, 0xcb, 0xb6, 0x40, 0x68, 0x37, 0xbf, 0x51, 0xf5
-};
-static uint8_t const C_cx_secp256r1_n[] = {
-  // n:  0xffffffff00000000ffffffffffffffffbce6faada7179e84f3b9cac2fc632551
-  0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xbc, 0xe6, 0xfa, 0xad, 0xa7, 0x17,
-  0x9e, 0x84, 0xf3, 0xb9, 0xca, 0xc2, 0xfc, 0x63, 0x25, 0x51
-};
-static uint8_t const C_cx_secp256r1_Hn[] = {
-  // Hn: 0x66e12d94f3d956202845b2392b6bec594699799c49bd6fa683244c95be79eea2
-  0x66, 0xe1, 0x2d, 0x94, 0xf3, 0xd9, 0x56, 0x20, 0x28, 0x45, 0xb2,
-  0x39, 0x2b, 0x6b, 0xec, 0x59, 0x46, 0x99, 0x79, 0x9c, 0x49, 0xbd,
-  0x6f, 0xa6, 0x83, 0x24, 0x4c, 0x95, 0xbe, 0x79, 0xee, 0xa2
-};
-
-#define C_cx_secp256r1_h C_cofactor_1
-
-cx_curve_weierstrass_t const C_cx_secp256r1 = {
-  .curve = CX_CURVE_SECP256R1,
-  .bit_size = 256,
-  .length = 32,
-
-  .a = C_cx_secp256r1_a,
-  .b = C_cx_secp256r1_b,
-  .p = C_cx_secp256r1_p,
-  .Gx = C_cx_secp256r1_Gx,
-  .Gy = C_cx_secp256r1_Gy,
-  .n = C_cx_secp256r1_n,
-  .h = C_cx_secp256r1_h,
-  .Hp = C_cx_secp256r1_Hp,
-  .Hn = C_cx_secp256r1_Hn,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                       secp384r1                                  --- */
-/* ------------------------------------------------------------------------ */
-
-static uint8_t const C_cx_secp384r1_a[] = {
-  // a:
-  // 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEFFFFFFFF0000000000000000FFFFFFFC
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0xff, 0xff, 0xff, 0xff,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xfc
-};
-static uint8_t const C_cx_secp384r1_b[] = {
-  // b:
-  0xB3, 0x31, 0x2F, 0xA7, 0xE2, 0x3E, 0xE7, 0xE4, 0x98, 0x8E, 0x05, 0x6B,
-  0xE3, 0xF8, 0x2D, 0x19, 0x18, 0x1D, 0x9C, 0x6E, 0xFE, 0x81, 0x41, 0x12,
-  0x03, 0x14, 0x08, 0x8F, 0x50, 0x13, 0x87, 0x5A, 0xC6, 0x56, 0x39, 0x8D,
-  0x8A, 0x2E, 0xD1, 0x9D, 0x2A, 0x85, 0xC8, 0xED, 0xD3, 0xEC, 0x2A, 0xEF
-};
-static uint8_t const C_cx_secp384r1_p[] = {
-  // p:
-  // 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEFFFFFFFF0000000000000000FFFFFFFF
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0xff, 0xff, 0xff, 0xff,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff
-};
-static uint8_t const C_cx_secp384r1_Hp[] = {
-  // Hp:
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00,
-  0xff, 0xff, 0xff, 0xfe, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02,
-  0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xfe, 0x00, 0x00, 0x00, 0x01
-};
-static uint8_t const C_cx_secp384r1_Gx[] = {
-  0xAA, 0x87, 0xCA, 0x22, 0xBE, 0x8B, 0x05, 0x37, 0x8E, 0xB1, 0xC7, 0x1E,
-  0xF3, 0x20, 0xAD, 0x74, 0x6E, 0x1D, 0x3B, 0x62, 0x8B, 0xA7, 0x9B, 0x98,
-  0x59, 0xF7, 0x41, 0xE0, 0x82, 0x54, 0x2A, 0x38, 0x55, 0x02, 0xF2, 0x5D,
-  0xBF, 0x55, 0x29, 0x6C, 0x3A, 0x54, 0x5E, 0x38, 0x72, 0x76, 0x0A, 0xB7
-};
-static uint8_t const C_cx_secp384r1_Gy[] = {
-  0x36, 0x17, 0xDE, 0x4A, 0x96, 0x26, 0x2C, 0x6F, 0x5D, 0x9E, 0x98, 0xBF,
-  0x92, 0x92, 0xDC, 0x29, 0xF8, 0xF4, 0x1D, 0xBD, 0x28, 0x9A, 0x14, 0x7C,
-  0xE9, 0xDA, 0x31, 0x13, 0xB5, 0xF0, 0xB8, 0xC0, 0x0A, 0x60, 0xB1, 0xCE,
-  0x1D, 0x7E, 0x81, 0x9D, 0x7A, 0x43, 0x1D, 0x7C, 0x90, 0xEA, 0x0E, 0x5F
-};
-static uint8_t const C_cx_secp384r1_n[] = {
-  // n:
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xC7, 0x63, 0x4D, 0x81, 0xF4, 0x37, 0x2D, 0xDF, 0x58, 0x1A, 0x0D, 0xB2,
-  0x48, 0xB0, 0xA7, 0x7A, 0xEC, 0xEC, 0x19, 0x6A, 0xCC, 0xC5, 0x29, 0x73
-};
-static uint8_t const C_cx_secp384r1_Hn[] = {
-  // Hn:
-  0x0c, 0x84, 0xee, 0x01, 0x2b, 0x39, 0xbf, 0x21, 0x3f, 0xb0, 0x5b, 0x7a,
-  0x28, 0x26, 0x68, 0x95, 0xd4, 0x0d, 0x49, 0x17, 0x4a, 0xab, 0x1c, 0xc5,
-  0xbc, 0x3e, 0x48, 0x3a, 0xfc, 0xb8, 0x29, 0x47, 0xff, 0x3d, 0x81, 0xe5,
-  0xdf, 0x1a, 0xa4, 0x19, 0x2d, 0x31, 0x9b, 0x24, 0x19, 0xb4, 0x09, 0xa9
-};
-
-#define C_cx_secp384r1_h C_cofactor_1
-
-cx_curve_weierstrass_t const C_cx_secp384r1 = {
-  .curve = CX_CURVE_SECP384R1,
-  .bit_size = 384,
-  .length = 48,
-
-  .a = C_cx_secp384r1_a,
-  .b = C_cx_secp384r1_b,
-  .p = C_cx_secp384r1_p,
-  .Gx = C_cx_secp384r1_Gx,
-  .Gy = C_cx_secp384r1_Gy,
-  .n = C_cx_secp384r1_n,
-  .h = C_cx_secp384r1_h,
-  .Hp = C_cx_secp384r1_Hp,
-  .Hn = C_cx_secp384r1_Hn,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                            secp521r1                             --- */
-/* ------------------------------------------------------------------------ */
-
-static uint8_t const C_cx_secp521r1_a[] = {
-  // a:
-  // 0x01FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFC
-  0x01, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFC
-};
-static uint8_t const C_cx_secp521r1_b[] = {
-  0x00, 0x51, 0x95, 0x3E, 0xB9, 0x61, 0x8E, 0x1C, 0x9A, 0x1F, 0x92,
-  0x9A, 0x21, 0xA0, 0xB6, 0x85, 0x40, 0xEE, 0xA2, 0xDA, 0x72, 0x5B,
-  0x99, 0xB3, 0x15, 0xF3, 0xB8, 0xB4, 0x89, 0x91, 0x8E, 0xF1, 0x09,
-  0xE1, 0x56, 0x19, 0x39, 0x51, 0xEC, 0x7E, 0x93, 0x7B, 0x16, 0x52,
-  0xC0, 0xBD, 0x3B, 0xB1, 0xBF, 0x07, 0x35, 0x73, 0xDF, 0x88, 0x3D,
-  0x2C, 0x34, 0xF1, 0xEF, 0x45, 0x1F, 0xD4, 0x6B, 0x50, 0x3F, 0x00
-};
-static uint8_t const C_cx_secp521r1_p[] = {
-  // p:
-  // 0x01FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF
-  0x01, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF
-};
-static uint8_t const C_cx_secp521r1_Hp[] = {
-  // Hp:
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-};
-static uint8_t const C_cx_secp521r1_Gx[] = {
-  0x00, 0xC6, 0x85, 0x8E, 0x06, 0xB7, 0x04, 0x04, 0xE9, 0xCD, 0x9E,
-  0x3E, 0xCB, 0x66, 0x23, 0x95, 0xB4, 0x42, 0x9C, 0x64, 0x81, 0x39,
-  0x05, 0x3F, 0xB5, 0x21, 0xF8, 0x28, 0xAF, 0x60, 0x6B, 0x4D, 0x3D,
-  0xBA, 0xA1, 0x4B, 0x5E, 0x77, 0xEF, 0xE7, 0x59, 0x28, 0xFE, 0x1D,
-  0xC1, 0x27, 0xA2, 0xFF, 0xA8, 0xDE, 0x33, 0x48, 0xB3, 0xC1, 0x85,
-  0x6A, 0x42, 0x9B, 0xF9, 0x7E, 0x7E, 0x31, 0xC2, 0xE5, 0xBD, 0x66
-};
-static uint8_t const C_cx_secp521r1_Gy[] = {
-  0x01, 0x18, 0x39, 0x29, 0x6A, 0x78, 0x9A, 0x3B, 0xC0, 0x04, 0x5C,
-  0x8A, 0x5F, 0xB4, 0x2C, 0x7D, 0x1B, 0xD9, 0x98, 0xF5, 0x44, 0x49,
-  0x57, 0x9B, 0x44, 0x68, 0x17, 0xAF, 0xBD, 0x17, 0x27, 0x3E, 0x66,
-  0x2C, 0x97, 0xEE, 0x72, 0x99, 0x5E, 0xF4, 0x26, 0x40, 0xC5, 0x50,
-  0xB9, 0x01, 0x3F, 0xAD, 0x07, 0x61, 0x35, 0x3C, 0x70, 0x86, 0xA2,
-  0x72, 0xC2, 0x40, 0x88, 0xBE, 0x94, 0x76, 0x9F, 0xD1, 0x66, 0x50
-};
-static uint8_t const C_cx_secp521r1_n[] = {
-  // n:
-  // 0x01FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFA51868783BF2F966B7FCC0148F709A5D03BB5C9B8899C47AEBB6FB71E91386409
-  0x01, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-  0xFA, 0x51, 0x86, 0x87, 0x83, 0xBF, 0x2F, 0x96, 0x6B, 0x7F, 0xCC,
-  0x01, 0x48, 0xF7, 0x09, 0xA5, 0xD0, 0x3B, 0xB5, 0xC9, 0xB8, 0x89,
-  0x9C, 0x47, 0xAE, 0xBB, 0x6F, 0xB7, 0x1E, 0x91, 0x38, 0x64, 0x09
-};
-static uint8_t const C_cx_secp521r1_Hn[] = {
-  // Hn:
-  0x01, 0xbc, 0xcf, 0xf3, 0xd1, 0x42, 0xb7, 0x75, 0x6e, 0x3e, 0xdd,
-  0x6e, 0x23, 0xd8, 0x2e, 0x49, 0xc7, 0xdc, 0x81, 0x3d, 0x00, 0xeb,
-  0xc5, 0x45, 0x05, 0x9b, 0x80, 0x5c, 0xc3, 0x49, 0x42, 0x90, 0xa7,
-  0xc8, 0x18, 0x47, 0x42, 0xd7, 0x00, 0xe0, 0x81, 0x09, 0x4e, 0x89,
-  0x3c, 0x9c, 0xf6, 0x52, 0x22, 0x7d, 0x46, 0x5b, 0x9f, 0x72, 0x66,
-  0xb4, 0x7e, 0xe7, 0xf0, 0x56, 0x21, 0x96, 0x51, 0x92, 0x0b, 0xb5
-};
-
-#define C_cx_secp521r1_h C_cofactor_1
-
-cx_curve_weierstrass_t const C_cx_secp521r1 = {
-  .curve = CX_CURVE_SECP521R1,
-  .bit_size = 521,
-  .length = 66,
-
-  .a = C_cx_secp521r1_a,
-  .b = C_cx_secp521r1_b,
-  .p = C_cx_secp521r1_p,
-  .Gx = C_cx_secp521r1_Gx,
-  .Gy = C_cx_secp521r1_Gy,
-  .n = C_cx_secp521r1_n,
-  .h = C_cx_secp521r1_h,
-  .Hn = C_cx_secp521r1_Hn,
-  .Hp = C_cx_secp521r1_Hp,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                            BrainpoolP256r1                       --- */
-/* ------------------------------------------------------------------------ */
-static uint8_t const C_cx_BrainpoolP256r1_a[] = {
-  // 0x7D5A0975FC2C3057EEF67530417AFFE7FB8055C126DC5C6CE94A4B44F330B5D9
-  0x7d, 0x5a, 0x09, 0x75, 0xfc, 0x2c, 0x30, 0x57, 0xee, 0xf6, 0x75,
-  0x30, 0x41, 0x7a, 0xff, 0xe7, 0xfb, 0x80, 0x55, 0xc1, 0x26, 0xdc,
-  0x5c, 0x6c, 0xe9, 0x4a, 0x4b, 0x44, 0xf3, 0x30, 0xb5, 0xd9
-};
-static uint8_t const C_cx_BrainpoolP256r1_b[] = {
-  // 0x26DC5C6CE94A4B44F330B5D9BBD77CBF958416295CF7E1CE6BCCDC18FF8C07B6
-  0x26, 0xdc, 0x5c, 0x6c, 0xe9, 0x4a, 0x4b, 0x44, 0xf3, 0x30, 0xb5,
-  0xd9, 0xbb, 0xd7, 0x7c, 0xbf, 0x95, 0x84, 0x16, 0x29, 0x5c, 0xf7,
-  0xe1, 0xce, 0x6b, 0xcc, 0xdc, 0x18, 0xff, 0x8c, 0x07, 0xb6
-};
-static uint8_t const C_cx_BrainpoolP256r1_p[] = {
-  // 0xA9FB57DBA1EEA9BC3E660A909D838D726E3BF623D52620282013481D1F6E5377
-  0xa9, 0xfb, 0x57, 0xdb, 0xa1, 0xee, 0xa9, 0xbc, 0x3e, 0x66, 0x0a,
-  0x90, 0x9d, 0x83, 0x8d, 0x72, 0x6e, 0x3b, 0xf6, 0x23, 0xd5, 0x26,
-  0x20, 0x28, 0x20, 0x13, 0x48, 0x1d, 0x1f, 0x6e, 0x53, 0x77
-};
-static uint8_t const C_cx_BrainpoolP256r1_Hp[] = {
-  0x47, 0x17, 0xaa, 0x21, 0xe5, 0x95, 0x7f, 0xa8, 0xa1, 0xec, 0xda,
-  0xcd, 0x6b, 0x1a, 0xc8, 0x07, 0x5c, 0xce, 0x4c, 0x26, 0x61, 0x4d,
-  0x4f, 0x4d, 0x8c, 0xfe, 0xdf, 0x7b, 0xa6, 0x46, 0x5b, 0x6c
-};
-static uint8_t const C_cx_BrainpoolP256r1_Gx[] = {
-  // 0x8BD2AEB9CB7E57CB2C4B482FFC81B7AFB9DE27E1E3BD23C23A4453BD9ACE3262
-  0x8b, 0xd2, 0xae, 0xb9, 0xcb, 0x7e, 0x57, 0xcb, 0x2c, 0x4b, 0x48,
-  0x2f, 0xfc, 0x81, 0xb7, 0xaf, 0xb9, 0xde, 0x27, 0xe1, 0xe3, 0xbd,
-  0x23, 0xc2, 0x3a, 0x44, 0x53, 0xbd, 0x9a, 0xce, 0x32, 0x62
-};
-static uint8_t const C_cx_BrainpoolP256r1_Gy[] = {
-  // 0x547EF835C3DAC4FD97F8461A14611DC9C27745132DED8E545C1D54C72F046997
-  0x54, 0x7e, 0xf8, 0x35, 0xc3, 0xda, 0xc4, 0xfd, 0x97, 0xf8, 0x46,
-  0x1a, 0x14, 0x61, 0x1d, 0xc9, 0xc2, 0x77, 0x45, 0x13, 0x2d, 0xed,
-  0x8e, 0x54, 0x5c, 0x1d, 0x54, 0xc7, 0x2f, 0x04, 0x69, 0x97
-};
-static uint8_t const C_cx_BrainpoolP256r1_n[] = {
-  // 0xA9FB57DBA1EEA9BC3E660A909D838D718C397AA3B561A6F7901E0E82974856A7
-  0xa9, 0xfb, 0x57, 0xdb, 0xa1, 0xee, 0xa9, 0xbc, 0x3e, 0x66, 0x0a,
-  0x90, 0x9d, 0x83, 0x8d, 0x71, 0x8c, 0x39, 0x7a, 0xa3, 0xb5, 0x61,
-  0xa6, 0xf7, 0x90, 0x1e, 0x0e, 0x82, 0x97, 0x48, 0x56, 0xa7
-};
-static uint8_t const C_cx_BrainpoolP256r1_Hn[] = {
-  0x0b, 0x25, 0xf1, 0xb9, 0xc3, 0x23, 0x67, 0x62, 0x9b, 0x7f, 0x25,
-  0xe7, 0x6c, 0x81, 0x5c, 0xb0, 0xf3, 0x5d, 0x17, 0x6a, 0x11, 0x34,
-  0xe4, 0xa0, 0xe1, 0xd8, 0xd8, 0xde, 0x33, 0x12, 0xfc, 0xa6
-};
-
-#define C_cx_BrainpoolP256r1_h C_cofactor_1
-cx_curve_weierstrass_t const C_cx_BrainpoolP256r1 = {
-  .curve = CX_CURVE_BrainPoolP256R1,
-  .bit_size = 256,
-  .length = 32,
-
-  .a = C_cx_BrainpoolP256r1_a,
-  .b = C_cx_BrainpoolP256r1_b,
-  .p = C_cx_BrainpoolP256r1_p,
-  .Gx = C_cx_BrainpoolP256r1_Gx,
-  .Gy = C_cx_BrainpoolP256r1_Gy,
-  .n = C_cx_BrainpoolP256r1_n,
-  .h = C_cx_BrainpoolP256r1_h,
-  .Hn = C_cx_BrainpoolP256r1_Hn,
-  .Hp = C_cx_BrainpoolP256r1_Hp,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                       BrainpoolP256t1                            --- */
-/* ------------------------------------------------------------------------ */
-static uint8_t const C_cx_BrainpoolP256t1_a[] = {
-  // 0xA9FB57DBA1EEA9BC3E660A909D838D726E3BF623D52620282013481D1F6E5374
-  0xa9, 0xfb, 0x57, 0xdb, 0xa1, 0xee, 0xa9, 0xbc, 0x3e, 0x66, 0x0a,
-  0x90, 0x9d, 0x83, 0x8d, 0x72, 0x6e, 0x3b, 0xf6, 0x23, 0xd5, 0x26,
-  0x20, 0x28, 0x20, 0x13, 0x48, 0x1d, 0x1f, 0x6e, 0x53, 0x74
-};
-static uint8_t const C_cx_BrainpoolP256t1_b[] = {
-  // 0x662C61C430D84EA4FE66A7733D0B76B7BF93EBC4AF2F49256AE58101FEE92B04
-  0x66, 0x2c, 0x61, 0xc4, 0x30, 0xd8, 0x4e, 0xa4, 0xfe, 0x66, 0xa7,
-  0x73, 0x3d, 0x0b, 0x76, 0xb7, 0xbf, 0x93, 0xeb, 0xc4, 0xaf, 0x2f,
-  0x49, 0x25, 0x6a, 0xe5, 0x81, 0x01, 0xfe, 0xe9, 0x2b, 0x04
-};
-static uint8_t const C_cx_BrainpoolP256t1_p[] = {
-  // 0xA9FB57DBA1EEA9BC3E660A909D838D726E3BF623D52620282013481D1F6E5377
-  0xa9, 0xfb, 0x57, 0xdb, 0xa1, 0xee, 0xa9, 0xbc, 0x3e, 0x66, 0x0a,
-  0x90, 0x9d, 0x83, 0x8d, 0x72, 0x6e, 0x3b, 0xf6, 0x23, 0xd5, 0x26,
-  0x20, 0x28, 0x20, 0x13, 0x48, 0x1d, 0x1f, 0x6e, 0x53, 0x77
-};
-static uint8_t const C_cx_BrainpoolP256t1_Hp[] = {
-  0x47, 0x17, 0xaa, 0x21, 0xe5, 0x95, 0x7f, 0xa8, 0xa1, 0xec, 0xda,
-  0xcd, 0x6b, 0x1a, 0xc8, 0x07, 0x5c, 0xce, 0x4c, 0x26, 0x61, 0x4d,
-  0x4f, 0x4d, 0x8c, 0xfe, 0xdf, 0x7b, 0xa6, 0x46, 0x5b, 0x6c
-};
-static uint8_t const C_cx_BrainpoolP256t1_Gx[] = {
-  // 0xA3E8EB3CC1CFE7B7732213B23A656149AFA142C47AAFBC2B79A191562E1305F4
-  0xa3, 0xe8, 0xeb, 0x3c, 0xc1, 0xcf, 0xe7, 0xb7, 0x73, 0x22, 0x13,
-  0xb2, 0x3a, 0x65, 0x61, 0x49, 0xaf, 0xa1, 0x42, 0xc4, 0x7a, 0xaf,
-  0xbc, 0x2b, 0x79, 0xa1, 0x91, 0x56, 0x2e, 0x13, 0x05, 0xf4
-};
-static uint8_t const C_cx_BrainpoolP256t1_Gy[] = {
-  // 0x2D996C823439C56D7F7B22E14644417E69BCB6DE39D027001DABE8F35B25C9BE
-  0x2d, 0x99, 0x6c, 0x82, 0x34, 0x39, 0xc5, 0x6d, 0x7f, 0x7b, 0x22,
-  0xe1, 0x46, 0x44, 0x41, 0x7e, 0x69, 0xbc, 0xb6, 0xde, 0x39, 0xd0,
-  0x27, 0x00, 0x1d, 0xab, 0xe8, 0xf3, 0x5b, 0x25, 0xc9, 0xbe
-};
-static uint8_t const C_cx_BrainpoolP256t1_n[] = {
-  // 0x
-  0xa9, 0xfb, 0x57, 0xdb, 0xa1, 0xee, 0xa9, 0xbc, 0x3e, 0x66, 0x0a,
-  0x90, 0x9d, 0x83, 0x8d, 0x71, 0x8c, 0x39, 0x7a, 0xa3, 0xb5, 0x61,
-  0xa6, 0xf7, 0x90, 0x1e, 0x0e, 0x82, 0x97, 0x48, 0x56, 0xa7
-};
-static uint8_t const C_cx_BrainpoolP256t1_Hn[] = {
-  0x0b, 0x25, 0xf1, 0xb9, 0xc3, 0x23, 0x67, 0x62, 0x9b, 0x7f, 0x25,
-  0xe7, 0x6c, 0x81, 0x5c, 0xb0, 0xf3, 0x5d, 0x17, 0x6a, 0x11, 0x34,
-  0xe4, 0xa0, 0xe1, 0xd8, 0xd8, 0xde, 0x33, 0x12, 0xfc, 0xa6
-};
-
-#define C_cx_BrainpoolP256t1_h C_cofactor_1
-cx_curve_weierstrass_t const C_cx_BrainpoolP256t1 = {
-  .curve = CX_CURVE_BrainPoolP256T1,
-  .bit_size = 256,
-  .length = 32,
-
-  .a = C_cx_BrainpoolP256t1_a,
-  .b = C_cx_BrainpoolP256t1_b,
-  .p = C_cx_BrainpoolP256t1_p,
-  .Gx = C_cx_BrainpoolP256t1_Gx,
-  .Gy = C_cx_BrainpoolP256t1_Gy,
-  .n = C_cx_BrainpoolP256t1_n,
-  .h = C_cx_BrainpoolP256t1_h,
-  .Hn = C_cx_BrainpoolP256t1_Hn,
-  .Hp = C_cx_BrainpoolP256t1_Hp,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                       BrainpoolP320r1                            --- */
-/* ------------------------------------------------------------------------ */
-static uint8_t const C_cx_BrainpoolP320r1_a[] = {
-  // a:  0x
-  0x3E, 0xE3, 0x0B, 0x56, 0x8F, 0xBA, 0xB0, 0xF8, 0x83, 0xCC,
-  0xEB, 0xD4, 0x6D, 0x3F, 0x3B, 0xB8, 0xA2, 0xA7, 0x35, 0x13,
-  0xF5, 0xEB, 0x79, 0xDA, 0x66, 0x19, 0x0E, 0xB0, 0x85, 0xFF,
-  0xA9, 0xF4, 0x92, 0xF3, 0x75, 0xA9, 0x7D, 0x86, 0x0E, 0xB4
-};
-static uint8_t const C_cx_BrainpoolP320r1_b[] = {
-  // b:  0x
-  0x52, 0x08, 0x83, 0x94, 0x9D, 0xFD, 0xBC, 0x42, 0xD3, 0xAD,
-  0x19, 0x86, 0x40, 0x68, 0x8A, 0x6F, 0xE1, 0x3F, 0x41, 0x34,
-  0x95, 0x54, 0xB4, 0x9A, 0xCC, 0x31, 0xDC, 0xCD, 0x88, 0x45,
-  0x39, 0x81, 0x6F, 0x5E, 0xB4, 0xAC, 0x8F, 0xB1, 0xF1, 0xA6
-};
-static uint8_t const C_cx_BrainpoolP320r1_p[] = {
-  // p:
-  // 0xD35E472036BC4FB7E13C785ED201E065F98FCFA6F6F40DEF4F92B9EC7893EC28FCD412B1F1B32E27
-  0xD3, 0x5E, 0x47, 0x20, 0x36, 0xBC, 0x4F, 0xB7, 0xE1, 0x3C,
-  0x78, 0x5E, 0xD2, 0x01, 0xE0, 0x65, 0xF9, 0x8F, 0xCF, 0xA6,
-  0xF6, 0xF4, 0x0D, 0xEF, 0x4F, 0x92, 0xB9, 0xEC, 0x78, 0x93,
-  0xEC, 0x28, 0xFC, 0xD4, 0x12, 0xB1, 0xF1, 0xB3, 0x2E, 0x27
-};
-static uint8_t const C_cx_BrainpoolP320r1_Hp[] = {
-  0x40, 0x3a, 0x05, 0x6e, 0x97, 0xa2, 0xc0, 0x9c, 0xdb, 0xc2,
-  0x1a, 0x81, 0x77, 0x27, 0xa6, 0xd4, 0xc5, 0x3a, 0xa9, 0x67,
-  0x5c, 0x40, 0x00, 0xf4, 0x10, 0x26, 0x0b, 0x24, 0xef, 0x82,
-  0x14, 0x48, 0xb4, 0xe0, 0x40, 0x50, 0xad, 0x8b, 0x40, 0x0d,
-};
-static uint8_t const C_cx_BrainpoolP320r1_Gx[] = {
-  // Gx:
-  // 0x43BD7E9AFB53D8B85289BCC48EE5BFE6F20137D10A087EB6E7871E2A10A599C710AF8D0D39E20611
-  0x43, 0xBD, 0x7E, 0x9A, 0xFB, 0x53, 0xD8, 0xB8, 0x52, 0x89,
-  0xBC, 0xC4, 0x8E, 0xE5, 0xBF, 0xE6, 0xF2, 0x01, 0x37, 0xD1,
-  0x0A, 0x08, 0x7E, 0xB6, 0xE7, 0x87, 0x1E, 0x2A, 0x10, 0xA5,
-  0x99, 0xC7, 0x10, 0xAF, 0x8D, 0x0D, 0x39, 0xE2, 0x06, 0x11
-};
-static uint8_t const C_cx_BrainpoolP320r1_Gy[] = {
-  // Gy,
-  // 0x14FDD05545EC1CC8AB4093247F77275E0743FFED117182EAA9C77877AAAC6AC7D35245D1692E8EE1
-  0x14, 0xFD, 0xD0, 0x55, 0x45, 0xEC, 0x1C, 0xC8, 0xAB, 0x40,
-  0x93, 0x24, 0x7F, 0x77, 0x27, 0x5E, 0x07, 0x43, 0xFF, 0xED,
-  0x11, 0x71, 0x82, 0xEA, 0xA9, 0xC7, 0x78, 0x77, 0xAA, 0xAC,
-  0x6A, 0xC7, 0xD3, 0x52, 0x45, 0xD1, 0x69, 0x2E, 0x8E, 0xE1
-};
-static uint8_t const C_cx_BrainpoolP320r1_n[] = {
-  // n:
-  // 0xD35E472036BC4FB7E13C785ED201E065F98FCFA5B68F12A32D482EC7EE8658E98691555B44C59311
-  0xD3, 0x5E, 0x47, 0x20, 0x36, 0xBC, 0x4F, 0xB7, 0xE1, 0x3C,
-  0x78, 0x5E, 0xD2, 0x01, 0xE0, 0x65, 0xF9, 0x8F, 0xCF, 0xA5,
-  0xB6, 0x8F, 0x12, 0xA3, 0x2D, 0x48, 0x2E, 0xC7, 0xEE, 0x86,
-  0x58, 0xE9, 0x86, 0x91, 0x55, 0x5B, 0x44, 0xC5, 0x93, 0x11
-};
-static uint8_t const C_cx_BrainpoolP320r1_Hn[] = {
-  0xa7, 0x0d, 0x3d, 0xb6, 0xd2, 0x66, 0x8a, 0x95, 0x3b, 0x53,
-  0x96, 0xfe, 0xb0, 0x0b, 0x95, 0x14, 0x07, 0xad, 0x16, 0x66,
-  0x45, 0x8b, 0xcc, 0x15, 0xdd, 0x86, 0x02, 0x95, 0xe4, 0xa3,
-  0xf1, 0xa0, 0xb9, 0xb6, 0xac, 0x1d, 0x66, 0x16, 0x3f, 0x17,
-
-};
-
-#define C_cx_BrainpoolP320r1_h C_cofactor_1
-cx_curve_weierstrass_t const C_cx_BrainpoolP320r1 = {
-  .curve = CX_CURVE_BrainPoolP320R1,
-  // 384, 48,
-  .bit_size = 320,
-  .length = 40,
-
-  .a = C_cx_BrainpoolP320r1_a,
-  .b = C_cx_BrainpoolP320r1_b,
-  .p = C_cx_BrainpoolP320r1_p,
-  .Gx = C_cx_BrainpoolP320r1_Gx,
-  .Gy = C_cx_BrainpoolP320r1_Gy,
-  .n = C_cx_BrainpoolP320r1_n,
-  .h = C_cx_BrainpoolP320r1_h,
-  .Hp = C_cx_BrainpoolP320r1_Hp,
-  .Hn = C_cx_BrainpoolP320r1_Hn,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                       BrainpoolP320t1                            --- */
-/* ------------------------------------------------------------------------ */
-static uint8_t const C_cx_BrainpoolP320t1_a[] = {
-  // a:
-  // 0xD35E472036BC4FB7E13C785ED201E065F98FCFA6F6F40DEF4F92B9EC7893EC28FCD412B1F1B32E24
-  0xD3, 0x5E, 0x47, 0x20, 0x36, 0xBC, 0x4F, 0xB7, 0xE1, 0x3C,
-  0x78, 0x5E, 0xD2, 0x01, 0xE0, 0x65, 0xF9, 0x8F, 0xCF, 0xA6,
-  0xF6, 0xF4, 0x0D, 0xEF, 0x4F, 0x92, 0xB9, 0xEC, 0x78, 0x93,
-  0xEC, 0x28, 0xFC, 0xD4, 0x12, 0xB1, 0xF1, 0xB3, 0x2E, 0x24
-};
-static uint8_t const C_cx_BrainpoolP320t1_b[] = {
-  // b:
-  // 0xA7F561E038EB1ED560B3D147DB782013064C19F27ED27C6780AAF77FB8A547CEB5B4FEF422340353
-  0xA7, 0xF5, 0x61, 0xE0, 0x38, 0xEB, 0x1E, 0xD5, 0x60, 0xB3,
-  0xD1, 0x47, 0xDB, 0x78, 0x20, 0x13, 0x06, 0x4C, 0x19, 0xF2,
-  0x7E, 0xD2, 0x7C, 0x67, 0x80, 0xAA, 0xF7, 0x7F, 0xB8, 0xA5,
-  0x47, 0xCE, 0xB5, 0xB4, 0xFE, 0xF4, 0x22, 0x34, 0x03, 0x53
-};
-static uint8_t const C_cx_BrainpoolP320t1_p[] = {
-  // p:
-  // 0xD35E472036BC4FB7E13C785ED201E065F98FCFA6F6F40DEF4F92B9EC7893EC28FCD412B1F1B32E27
-  0xD3, 0x5E, 0x47, 0x20, 0x36, 0xBC, 0x4F, 0xB7, 0xE1, 0x3C,
-  0x78, 0x5E, 0xD2, 0x01, 0xE0, 0x65, 0xF9, 0x8F, 0xCF, 0xA6,
-  0xF6, 0xF4, 0x0D, 0xEF, 0x4F, 0x92, 0xB9, 0xEC, 0x78, 0x93,
-  0xEC, 0x28, 0xFC, 0xD4, 0x12, 0xB1, 0xF1, 0xB3, 0x2E, 0x27
-};
-static uint8_t const C_cx_BrainpoolP320t1_Hp[] = {
-  0x40, 0x3a, 0x05, 0x6e, 0x97, 0xa2, 0xc0, 0x9c, 0xdb, 0xc2,
-  0x1a, 0x81, 0x77, 0x27, 0xa6, 0xd4, 0xc5, 0x3a, 0xa9, 0x67,
-  0x5c, 0x40, 0x00, 0xf4, 0x10, 0x26, 0x0b, 0x24, 0xef, 0x82,
-  0x14, 0x48, 0xb4, 0xe0, 0x40, 0x50, 0xad, 0x8b, 0x40, 0x0d
-};
-static uint8_t const C_cx_BrainpoolP320t1_Gx[] = {
-  // Gx:
-  // 0x925BE9FB01AFC6FB4D3E7D4990010F813408AB106C4F09CB7EE07868CC136FFF3357F624A21BED52
-  0x92, 0x5B, 0xE9, 0xFB, 0x01, 0xAF, 0xC6, 0xFB, 0x4D, 0x3E,
-  0x7D, 0x49, 0x90, 0x01, 0x0F, 0x81, 0x34, 0x08, 0xAB, 0x10,
-  0x6C, 0x4F, 0x09, 0xCB, 0x7E, 0xE0, 0x78, 0x68, 0xCC, 0x13,
-  0x6F, 0xFF, 0x33, 0x57, 0xF6, 0x24, 0xA2, 0x1B, 0xED, 0x52
-};
-static uint8_t const C_cx_BrainpoolP320t1_Gy[] = {
-  // Gy:
-  // 0x63BA3A7A27483EBF6671DBEF7ABB30EBEE084E58A0B077AD42A5A0989D1EE71B1B9BC0455FB0D2C3
-  0x63, 0xBA, 0x3A, 0x7A, 0x27, 0x48, 0x3E, 0xBF, 0x66, 0x71,
-  0xDB, 0xEF, 0x7A, 0xBB, 0x30, 0xEB, 0xEE, 0x08, 0x4E, 0x58,
-  0xA0, 0xB0, 0x77, 0xAD, 0x42, 0xA5, 0xA0, 0x98, 0x9D, 0x1E,
-  0xE7, 0x1B, 0x1B, 0x9B, 0xC0, 0x45, 0x5F, 0xB0, 0xD2, 0xC3
-};
-static uint8_t const C_cx_BrainpoolP320t1_n[] = {
-  // n:
-  // 0xD35E472036BC4FB7E13C785ED201E065F98FCFA5B68F12A32D482EC7EE8658E98691555B44C59311
-  0xD3, 0x5E, 0x47, 0x20, 0x36, 0xBC, 0x4F, 0xB7, 0xE1, 0x3C,
-  0x78, 0x5E, 0xD2, 0x01, 0xE0, 0x65, 0xF9, 0x8F, 0xCF, 0xA5,
-  0xB6, 0x8F, 0x12, 0xA3, 0x2D, 0x48, 0x2E, 0xC7, 0xEE, 0x86,
-  0x58, 0xE9, 0x86, 0x91, 0x55, 0x5B, 0x44, 0xC5, 0x93, 0x11
-};
-static uint8_t const C_cx_BrainpoolP320t1_Hn[] = {
-  0xa7, 0x0d, 0x3d, 0xb6, 0xd2, 0x66, 0x8a, 0x95, 0x3b, 0x53,
-  0x96, 0xfe, 0xb0, 0x0b, 0x95, 0x14, 0x07, 0xad, 0x16, 0x66,
-  0x45, 0x8b, 0xcc, 0x15, 0xdd, 0x86, 0x02, 0x95, 0xe4, 0xa3,
-  0xf1, 0xa0, 0xb9, 0xb6, 0xac, 0x1d, 0x66, 0x16, 0x3f, 0x17
-};
-
-#define C_cx_BrainpoolP320t1_h C_cofactor_1
-cx_curve_weierstrass_t const C_cx_BrainpoolP320t1 = {
-  .curve = CX_CURVE_BrainPoolP320T1,
-  // 384, 48,
-  .bit_size = 320,
-  .length = 40,
-
-  .a = C_cx_BrainpoolP320t1_a,
-  .b = C_cx_BrainpoolP320t1_b,
-  .p = C_cx_BrainpoolP320t1_p,
-  .Gx = C_cx_BrainpoolP320t1_Gx,
-  .Gy = C_cx_BrainpoolP320t1_Gy,
-  .n = C_cx_BrainpoolP320t1_n,
-  .h = C_cx_BrainpoolP320t1_h,
-  .Hp = C_cx_BrainpoolP320t1_Hp,
-  .Hn = C_cx_BrainpoolP320t1_Hn,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                       BrainpoolP384r1                            --- */
-/* ------------------------------------------------------------------------ */
-static uint8_t const C_cx_BrainpoolP384r1_a[] = {
-  // a:
-  // 0x7BC382C63D8C150C3C72080ACE05AFA0C2BEA28E4FB22787139165EFBA91F90F8AA5814A503AD4EB04A8C7DD22CE2826
-  0x7B, 0xC3, 0x82, 0xC6, 0x3D, 0x8C, 0x15, 0x0C, 0x3C, 0x72, 0x08, 0x0A,
-  0xCE, 0x05, 0xAF, 0xA0, 0xC2, 0xBE, 0xA2, 0x8E, 0x4F, 0xB2, 0x27, 0x87,
-  0x13, 0x91, 0x65, 0xEF, 0xBA, 0x91, 0xF9, 0x0F, 0x8A, 0xA5, 0x81, 0x4A,
-  0x50, 0x3A, 0xD4, 0xEB, 0x04, 0xA8, 0xC7, 0xDD, 0x22, 0xCE, 0x28, 0x26
-};
-static uint8_t const C_cx_BrainpoolP384r1_b[] = {
-  // b:
-  // 0x04A8C7DD22CE28268B39B55416F0447C2FB77DE107DCD2A62E880EA53EEB62D57CB4390295DBC9943AB78696FA504C11
-  0x04, 0xA8, 0xC7, 0xDD, 0x22, 0xCE, 0x28, 0x26, 0x8B, 0x39, 0xB5, 0x54,
-  0x16, 0xF0, 0x44, 0x7C, 0x2F, 0xB7, 0x7D, 0xE1, 0x07, 0xDC, 0xD2, 0xA6,
-  0x2E, 0x88, 0x0E, 0xA5, 0x3E, 0xEB, 0x62, 0xD5, 0x7C, 0xB4, 0x39, 0x02,
-  0x95, 0xDB, 0xC9, 0x94, 0x3A, 0xB7, 0x86, 0x96, 0xFA, 0x50, 0x4C, 0x11
-};
-static uint8_t const C_cx_BrainpoolP384r1_p[] = {
-  // p:
-  // 0x8CB91E82A3386D280F5D6F7E50E641DF152F7109ED5456B412B1DA197FB71123ACD3A729901D1A71874700133107EC53
-  0x8C, 0xB9, 0x1E, 0x82, 0xA3, 0x38, 0x6D, 0x28, 0x0F, 0x5D, 0x6F, 0x7E,
-  0x50, 0xE6, 0x41, 0xDF, 0x15, 0x2F, 0x71, 0x09, 0xED, 0x54, 0x56, 0xB4,
-  0x12, 0xB1, 0xDA, 0x19, 0x7F, 0xB7, 0x11, 0x23, 0xAC, 0xD3, 0xA7, 0x29,
-  0x90, 0x1D, 0x1A, 0x71, 0x87, 0x47, 0x00, 0x13, 0x31, 0x07, 0xEC, 0x53
-};
-static uint8_t const C_cx_BrainpoolP384r1_Hp[] = {
-  0x36, 0xbf, 0x68, 0x83, 0x17, 0x8d, 0xf8, 0x42, 0xd5, 0xc6, 0xef, 0x3b,
-  0xa5, 0x7e, 0x05, 0x2c, 0x62, 0x14, 0x01, 0x91, 0x99, 0x18, 0xd5, 0xaf,
-  0x8e, 0x28, 0xf9, 0x9c, 0xc9, 0x94, 0x08, 0x99, 0x53, 0x52, 0x83, 0x34,
-  0x3d, 0x7f, 0xd9, 0x65, 0x08, 0x7c, 0xef, 0xff, 0x40, 0xb6, 0x4b, 0xde
-};
-static uint8_t const C_cx_BrainpoolP384r1_Gx[] = {
-  // Gx:
-  // 0x1D1C64F068CF45FFA2A63A81B7C13F6B8847A3E77EF14FE3DB7FCAFE0CBD10E8E826E03436D646AAEF87B2E247D4AF1E
-  0x1D, 0x1C, 0x64, 0xF0, 0x68, 0xCF, 0x45, 0xFF, 0xA2, 0xA6, 0x3A, 0x81,
-  0xB7, 0xC1, 0x3F, 0x6B, 0x88, 0x47, 0xA3, 0xE7, 0x7E, 0xF1, 0x4F, 0xE3,
-  0xDB, 0x7F, 0xCA, 0xFE, 0x0C, 0xBD, 0x10, 0xE8, 0xE8, 0x26, 0xE0, 0x34,
-  0x36, 0xD6, 0x46, 0xAA, 0xEF, 0x87, 0xB2, 0xE2, 0x47, 0xD4, 0xAF, 0x1E
-};
-static uint8_t const C_cx_BrainpoolP384r1_Gy[] = {
-  // Gy:
-  // 0x8ABE1D7520F9C2A45CB1EB8E95CFD55262B70B29FEEC5864E19C054FF99129280E4646217791811142820341263C5315
-  0x8A, 0xBE, 0x1D, 0x75, 0x20, 0xF9, 0xC2, 0xA4, 0x5C, 0xB1, 0xEB, 0x8E,
-  0x95, 0xCF, 0xD5, 0x52, 0x62, 0xB7, 0x0B, 0x29, 0xFE, 0xEC, 0x58, 0x64,
-  0xE1, 0x9C, 0x05, 0x4F, 0xF9, 0x91, 0x29, 0x28, 0x0E, 0x46, 0x46, 0x21,
-  0x77, 0x91, 0x81, 0x11, 0x42, 0x82, 0x03, 0x41, 0x26, 0x3C, 0x53, 0x15
-};
-static uint8_t const C_cx_BrainpoolP384r1_n[] = {
-  // n:
-  // 0x8CB91E82A3386D280F5D6F7E50E641DF152F7109ED5456B31F166E6CAC0425A7CF3AB6AF6B7FC3103B883202E9046565
-  0x8C, 0xB9, 0x1E, 0x82, 0xA3, 0x38, 0x6D, 0x28, 0x0F, 0x5D, 0x6F, 0x7E,
-  0x50, 0xE6, 0x41, 0xDF, 0x15, 0x2F, 0x71, 0x09, 0xED, 0x54, 0x56, 0xB3,
-  0x1F, 0x16, 0x6E, 0x6C, 0xAC, 0x04, 0x25, 0xA7, 0xCF, 0x3A, 0xB6, 0xAF,
-  0x6B, 0x7F, 0xC3, 0x10, 0x3B, 0x88, 0x32, 0x02, 0xE9, 0x04, 0x65, 0x65
-};
-static uint8_t const C_cx_BrainpoolP384r1_Hn[] = {
-  0x0c, 0xe8, 0x94, 0x1a, 0x61, 0x4e, 0x97, 0xc2, 0x8f, 0x88, 0x6d, 0xc9,
-  0x65, 0x16, 0x5f, 0xdb, 0x57, 0x4a, 0x74, 0xcb, 0x52, 0xd7, 0x48, 0xff,
-  0x2a, 0x92, 0x7e, 0x3b, 0x98, 0x02, 0x68, 0x8a, 0x37, 0x26, 0x4e, 0x20,
-  0x2f, 0x2b, 0x6b, 0x6e, 0xac, 0x4e, 0xd3, 0xa2, 0xde, 0x77, 0x1c, 0x8e
-};
-
-#define C_cx_BrainpoolP384r1_h C_cofactor_1
-cx_curve_weierstrass_t const C_cx_BrainpoolP384r1 = {
-  .curve = CX_CURVE_BrainPoolP384R1,
-  .bit_size = 384,
-  .length = 48,
-
-  .a = C_cx_BrainpoolP384r1_a,
-  .b = C_cx_BrainpoolP384r1_b,
-  .p = C_cx_BrainpoolP384r1_p,
-  .Gx = C_cx_BrainpoolP384r1_Gx,
-  .Gy = C_cx_BrainpoolP384r1_Gy,
-  .n = C_cx_BrainpoolP384r1_n,
-  .h = C_cx_BrainpoolP384r1_h,
-  .Hp = C_cx_BrainpoolP384r1_Hp,
-  .Hn = C_cx_BrainpoolP384r1_Hn,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                       BrainpoolP384t1                            --- */
-/* ------------------------------------------------------------------------ */
-static uint8_t const C_cx_BrainpoolP384t1_a[] = {
-  // a:
-  // 0x8CB91E82A3386D280F5D6F7E50E641DF152F7109ED5456B412B1DA197FB71123ACD3A729901D1A71874700133107EC50
-  0x8C, 0xB9, 0x1E, 0x82, 0xA3, 0x38, 0x6D, 0x28, 0x0F, 0x5D, 0x6F, 0x7E,
-  0x50, 0xE6, 0x41, 0xDF, 0x15, 0x2F, 0x71, 0x09, 0xED, 0x54, 0x56, 0xB4,
-  0x12, 0xB1, 0xDA, 0x19, 0x7F, 0xB7, 0x11, 0x23, 0xAC, 0xD3, 0xA7, 0x29,
-  0x90, 0x1D, 0x1A, 0x71, 0x87, 0x47, 0x00, 0x13, 0x31, 0x07, 0xEC, 0x50
-};
-static uint8_t const C_cx_BrainpoolP384t1_b[] = {
-  // b:
-  // 0x7F519EADA7BDA81BD826DBA647910F8C4B9346ED8CCDC64E4B1ABD11756DCE1D2074AA263B88805CED70355A33B471EE
-  0x7F, 0x51, 0x9E, 0xAD, 0xA7, 0xBD, 0xA8, 0x1B, 0xD8, 0x26, 0xDB, 0xA6,
-  0x47, 0x91, 0x0F, 0x8C, 0x4B, 0x93, 0x46, 0xED, 0x8C, 0xCD, 0xC6, 0x4E,
-  0x4B, 0x1A, 0xBD, 0x11, 0x75, 0x6D, 0xCE, 0x1D, 0x20, 0x74, 0xAA, 0x26,
-  0x3B, 0x88, 0x80, 0x5C, 0xED, 0x70, 0x35, 0x5A, 0x33, 0xB4, 0x71, 0xEE
-};
-static uint8_t const C_cx_BrainpoolP384t1_p[] = {
-  // p:
-  // 0x8CB91E82A3386D280F5D6F7E50E641DF152F7109ED5456B412B1DA197FB71123ACD3A729901D1A71874700133107EC53
-  0x8C, 0xB9, 0x1E, 0x82, 0xA3, 0x38, 0x6D, 0x28, 0x0F, 0x5D, 0x6F, 0x7E,
-  0x50, 0xE6, 0x41, 0xDF, 0x15, 0x2F, 0x71, 0x09, 0xED, 0x54, 0x56, 0xB4,
-  0x12, 0xB1, 0xDA, 0x19, 0x7F, 0xB7, 0x11, 0x23, 0xAC, 0xD3, 0xA7, 0x29,
-  0x90, 0x1D, 0x1A, 0x71, 0x87, 0x47, 0x00, 0x13, 0x31, 0x07, 0xEC, 0x53
-};
-static uint8_t const C_cx_BrainpoolP384t1_Hp[] = {
-  0x36, 0xbf, 0x68, 0x83, 0x17, 0x8d, 0xf8, 0x42, 0xd5, 0xc6, 0xef, 0x3b,
-  0xa5, 0x7e, 0x05, 0x2c, 0x62, 0x14, 0x01, 0x91, 0x99, 0x18, 0xd5, 0xaf,
-  0x8e, 0x28, 0xf9, 0x9c, 0xc9, 0x94, 0x08, 0x99, 0x53, 0x52, 0x83, 0x34,
-  0x3d, 0x7f, 0xd9, 0x65, 0x08, 0x7c, 0xef, 0xff, 0x40, 0xb6, 0x4b, 0xde
-};
-static uint8_t const C_cx_BrainpoolP384t1_Gx[] = {
-  // Gx:
-  // 0x18DE98B02DB9A306F2AFCD7235F72A819B80AB12EBD653172476FECD462AABFFC4FF191B946A5F54D8D0AA2F418808CC
-  0x18, 0xDE, 0x98, 0xB0, 0x2D, 0xB9, 0xA3, 0x06, 0xF2, 0xAF, 0xCD, 0x72,
-  0x35, 0xF7, 0x2A, 0x81, 0x9B, 0x80, 0xAB, 0x12, 0xEB, 0xD6, 0x53, 0x17,
-  0x24, 0x76, 0xFE, 0xCD, 0x46, 0x2A, 0xAB, 0xFF, 0xC4, 0xFF, 0x19, 0x1B,
-  0x94, 0x6A, 0x5F, 0x54, 0xD8, 0xD0, 0xAA, 0x2F, 0x41, 0x88, 0x08, 0xCC
-};
-static uint8_t const C_cx_BrainpoolP384t1_Gy[] = {
-  // Gy:
-  // 0x25AB056962D30651A114AFD2755AD336747F93475B7A1FCA3B88F2B6A208CCFE469408584DC2B2912675BF5B9E582928
-  0x25, 0xAB, 0x05, 0x69, 0x62, 0xD3, 0x06, 0x51, 0xA1, 0x14, 0xAF, 0xD2,
-  0x75, 0x5A, 0xD3, 0x36, 0x74, 0x7F, 0x93, 0x47, 0x5B, 0x7A, 0x1F, 0xCA,
-  0x3B, 0x88, 0xF2, 0xB6, 0xA2, 0x08, 0xCC, 0xFE, 0x46, 0x94, 0x08, 0x58,
-  0x4D, 0xC2, 0xB2, 0x91, 0x26, 0x75, 0xBF, 0x5B, 0x9E, 0x58, 0x29, 0x28
-};
-static uint8_t const C_cx_BrainpoolP384t1_n[] = {
-  // n:
-  // 0x8CB91E82A3386D280F5D6F7E50E641DF152F7109ED5456B31F166E6CAC0425A7CF3AB6AF6B7FC3103B883202E9046565
-  0x8C, 0xB9, 0x1E, 0x82, 0xA3, 0x38, 0x6D, 0x28, 0x0F, 0x5D, 0x6F, 0x7E,
-  0x50, 0xE6, 0x41, 0xDF, 0x15, 0x2F, 0x71, 0x09, 0xED, 0x54, 0x56, 0xB3,
-  0x1F, 0x16, 0x6E, 0x6C, 0xAC, 0x04, 0x25, 0xA7, 0xCF, 0x3A, 0xB6, 0xAF,
-  0x6B, 0x7F, 0xC3, 0x10, 0x3B, 0x88, 0x32, 0x02, 0xE9, 0x04, 0x65, 0x65
-};
-static uint8_t const C_cx_BrainpoolP384t1_Hn[] = {
-  0x0c, 0xe8, 0x94, 0x1a, 0x61, 0x4e, 0x97, 0xc2, 0x8f, 0x88, 0x6d, 0xc9,
-  0x65, 0x16, 0x5f, 0xdb, 0x57, 0x4a, 0x74, 0xcb, 0x52, 0xd7, 0x48, 0xff,
-  0x2a, 0x92, 0x7e, 0x3b, 0x98, 0x02, 0x68, 0x8a, 0x37, 0x26, 0x4e, 0x20,
-  0x2f, 0x2b, 0x6b, 0x6e, 0xac, 0x4e, 0xd3, 0xa2, 0xde, 0x77, 0x1c, 0x8e
-};
-
-#define C_cx_BrainpoolP384t1_h C_cofactor_1
-cx_curve_weierstrass_t const C_cx_BrainpoolP384t1 = {
-  .curve = CX_CURVE_BrainPoolP384T1,
-  .bit_size = 384,
-  .length = 48,
-
-  .a = C_cx_BrainpoolP384t1_a,
-  .b = C_cx_BrainpoolP384t1_b,
-  .p = C_cx_BrainpoolP384t1_p,
-  .Gx = C_cx_BrainpoolP384t1_Gx,
-  .Gy = C_cx_BrainpoolP384t1_Gy,
-  .n = C_cx_BrainpoolP384t1_n,
-  .h = C_cx_BrainpoolP384t1_h,
-  .Hp = C_cx_BrainpoolP384t1_Hp,
-  .Hn = C_cx_BrainpoolP384t1_Hn,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                       BrainpoolP512r1                            --- */
-/* ------------------------------------------------------------------------ */
-static uint8_t const C_cx_BrainpoolP512r1_a[] = {
-  // a:
-  // 0x7830A3318B603B89E2327145AC234CC594CBDD8D3DF91610A83441CAEA9863BC2DED5D5AA8253AA10A2EF1C98B9AC8B57F1117A72BF2C7B9E7C1AC4D77FC94CA
-  0x78, 0x30, 0xA3, 0x31, 0x8B, 0x60, 0x3B, 0x89, 0xE2, 0x32, 0x71, 0x45, 0xAC,
-  0x23, 0x4C, 0xC5, 0x94, 0xCB, 0xDD, 0x8D, 0x3D, 0xF9, 0x16, 0x10, 0xA8, 0x34,
-  0x41, 0xCA, 0xEA, 0x98, 0x63, 0xBC, 0x2D, 0xED, 0x5D, 0x5A, 0xA8, 0x25, 0x3A,
-  0xA1, 0x0A, 0x2E, 0xF1, 0xC9, 0x8B, 0x9A, 0xC8, 0xB5, 0x7F, 0x11, 0x17, 0xA7,
-  0x2B, 0xF2, 0xC7, 0xB9, 0xE7, 0xC1, 0xAC, 0x4D, 0x77, 0xFC, 0x94, 0xCA
-};
-static uint8_t const C_cx_BrainpoolP512r1_b[] = {
-  // b:
-  // 0x3DF91610A83441CAEA9863BC2DED5D5AA8253AA10A2EF1C98B9AC8B57F1117A72BF2C7B9E7C1AC4D77FC94CADC083E67984050B75EBAE5DD2809BD638016F723
-  0x3D, 0xF9, 0x16, 0x10, 0xA8, 0x34, 0x41, 0xCA, 0xEA, 0x98, 0x63, 0xBC, 0x2D,
-  0xED, 0x5D, 0x5A, 0xA8, 0x25, 0x3A, 0xA1, 0x0A, 0x2E, 0xF1, 0xC9, 0x8B, 0x9A,
-  0xC8, 0xB5, 0x7F, 0x11, 0x17, 0xA7, 0x2B, 0xF2, 0xC7, 0xB9, 0xE7, 0xC1, 0xAC,
-  0x4D, 0x77, 0xFC, 0x94, 0xCA, 0xDC, 0x08, 0x3E, 0x67, 0x98, 0x40, 0x50, 0xB7,
-  0x5E, 0xBA, 0xE5, 0xDD, 0x28, 0x09, 0xBD, 0x63, 0x80, 0x16, 0xF7, 0x23
-};
-static uint8_t const C_cx_BrainpoolP512r1_p[] = {
-  // p:
-  // 0xAADD9DB8DBE9C48B3FD4E6AE33C9FC07CB308DB3B3C9D20ED6639CCA703308717D4D9B009BC66842AECDA12AE6A380E62881FF2F2D82C68528AA6056583A48F3
-  0xAA, 0xDD, 0x9D, 0xB8, 0xDB, 0xE9, 0xC4, 0x8B, 0x3F, 0xD4, 0xE6, 0xAE, 0x33,
-  0xC9, 0xFC, 0x07, 0xCB, 0x30, 0x8D, 0xB3, 0xB3, 0xC9, 0xD2, 0x0E, 0xD6, 0x63,
-  0x9C, 0xCA, 0x70, 0x33, 0x08, 0x71, 0x7D, 0x4D, 0x9B, 0x00, 0x9B, 0xC6, 0x68,
-  0x42, 0xAE, 0xCD, 0xA1, 0x2A, 0xE6, 0xA3, 0x80, 0xE6, 0x28, 0x81, 0xFF, 0x2F,
-  0x2D, 0x82, 0xC6, 0x85, 0x28, 0xAA, 0x60, 0x56, 0x58, 0x3A, 0x48, 0xF3
-};
-static uint8_t const C_cx_BrainpoolP512r1_Hp[] = {
-  0x3c, 0x4c, 0x9d, 0x05, 0xa9, 0xff, 0x64, 0x50, 0x20, 0x2e, 0x19, 0x40, 0x20,
-  0x56, 0xee, 0xcc, 0xa1, 0x6d, 0xaa, 0x5f, 0xd4, 0x2b, 0xff, 0x83, 0x19, 0x48,
-  0x6f, 0xd8, 0xd5, 0x89, 0x80, 0x57, 0xe0, 0xc1, 0x9a, 0x77, 0x83, 0x51, 0x4a,
-  0x25, 0x53, 0xb7, 0xf9, 0xbc, 0x90, 0x5a, 0xff, 0xd3, 0x79, 0x3f, 0xb1, 0x30,
-  0x27, 0x15, 0x79, 0x05, 0x49, 0xad, 0x14, 0x4a, 0x61, 0x58, 0xf2, 0x05
-};
-static uint8_t const C_cx_BrainpoolP512r1_Gx[] = {
-  // Gx:
-  // 0x81AEE4BDD82ED9645A21322E9C4C6A9385ED9F70B5D916C1B43B62EEF4D0098EFF3B1F78E2D0D48D50D1687B93B97D5F7C6D5047406A5E688B352209BCB9F822
-  0x81, 0xAE, 0xE4, 0xBD, 0xD8, 0x2E, 0xD9, 0x64, 0x5A, 0x21, 0x32, 0x2E, 0x9C,
-  0x4C, 0x6A, 0x93, 0x85, 0xED, 0x9F, 0x70, 0xB5, 0xD9, 0x16, 0xC1, 0xB4, 0x3B,
-  0x62, 0xEE, 0xF4, 0xD0, 0x09, 0x8E, 0xFF, 0x3B, 0x1F, 0x78, 0xE2, 0xD0, 0xD4,
-  0x8D, 0x50, 0xD1, 0x68, 0x7B, 0x93, 0xB9, 0x7D, 0x5F, 0x7C, 0x6D, 0x50, 0x47,
-  0x40, 0x6A, 0x5E, 0x68, 0x8B, 0x35, 0x22, 0x09, 0xBC, 0xB9, 0xF8, 0x22
-};
-static uint8_t const C_cx_BrainpoolP512r1_Gy[] = {
-  // Gy:
-  // 0x7DDE385D566332ECC0EABFA9CF7822FDF209F70024A57B1AA000C55B881F8111B2DCDE494A5F485E5BCA4BD88A2763AED1CA2B2FA8F0540678CD1E0F3AD80892
-  0x7D, 0xDE, 0x38, 0x5D, 0x56, 0x63, 0x32, 0xEC, 0xC0, 0xEA, 0xBF, 0xA9, 0xCF,
-  0x78, 0x22, 0xFD, 0xF2, 0x09, 0xF7, 0x00, 0x24, 0xA5, 0x7B, 0x1A, 0xA0, 0x00,
-  0xC5, 0x5B, 0x88, 0x1F, 0x81, 0x11, 0xB2, 0xDC, 0xDE, 0x49, 0x4A, 0x5F, 0x48,
-  0x5E, 0x5B, 0xCA, 0x4B, 0xD8, 0x8A, 0x27, 0x63, 0xAE, 0xD1, 0xCA, 0x2B, 0x2F,
-  0xA8, 0xF0, 0x54, 0x06, 0x78, 0xCD, 0x1E, 0x0F, 0x3A, 0xD8, 0x08, 0x92
-};
-static uint8_t const C_cx_BrainpoolP512r1_n[] = {
-  // n:
-  // 0xAADD9DB8DBE9C48B3FD4E6AE33C9FC07CB308DB3B3C9D20ED6639CCA70330870553E5C414CA92619418661197FAC10471DB1D381085DDADDB58796829CA90069
-  0xAA, 0xDD, 0x9D, 0xB8, 0xDB, 0xE9, 0xC4, 0x8B, 0x3F, 0xD4, 0xE6, 0xAE, 0x33,
-  0xC9, 0xFC, 0x07, 0xCB, 0x30, 0x8D, 0xB3, 0xB3, 0xC9, 0xD2, 0x0E, 0xD6, 0x63,
-  0x9C, 0xCA, 0x70, 0x33, 0x08, 0x70, 0x55, 0x3E, 0x5C, 0x41, 0x4C, 0xA9, 0x26,
-  0x19, 0x41, 0x86, 0x61, 0x19, 0x7F, 0xAC, 0x10, 0x47, 0x1D, 0xB1, 0xD3, 0x81,
-  0x08, 0x5D, 0xDA, 0xDD, 0xB5, 0x87, 0x96, 0x82, 0x9C, 0xA9, 0x00, 0x69
-};
-static uint8_t const C_cx_BrainpoolP512r1_Hn[] = {
-  0xa7, 0x94, 0x58, 0x6a, 0x71, 0x84, 0x07, 0xb0, 0x95, 0xdf, 0x1b, 0x4c, 0x19,
-  0x4b, 0x2e, 0x56, 0x72, 0x3c, 0x37, 0xa2, 0x2f, 0x16, 0xbb, 0xdf, 0xd7, 0xf9,
-  0xcc, 0x26, 0x3b, 0x79, 0x0d, 0xe3, 0xa6, 0xf2, 0x30, 0xc7, 0x2f, 0x02, 0x07,
-  0xe8, 0x3e, 0xc6, 0x4b, 0xd0, 0x33, 0xb7, 0x62, 0x7f, 0x08, 0x86, 0xb7, 0x58,
-  0x95, 0x28, 0x3d, 0xdd, 0xd2, 0xa3, 0x68, 0x1e, 0xcd, 0xa8, 0x16, 0x71
-};
-
-#define C_cx_BrainpoolP512r1_h C_cofactor_1
-cx_curve_weierstrass_t const C_cx_BrainpoolP512r1 = {
-  .curve = CX_CURVE_BrainPoolP512R1,
-  .bit_size = 512,
-  .length = 64,
-
-  .a = C_cx_BrainpoolP512r1_a,
-  .b = C_cx_BrainpoolP512r1_b,
-  .p = C_cx_BrainpoolP512r1_p,
-  .Gx = C_cx_BrainpoolP512r1_Gx,
-  .Gy = C_cx_BrainpoolP512r1_Gy,
-  .n = C_cx_BrainpoolP512r1_n,
-  .h = C_cx_BrainpoolP512r1_h,
-  .Hp = C_cx_BrainpoolP512r1_Hp,
-  .Hn = C_cx_BrainpoolP512r1_Hn,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                       BrainpoolP512t1                            --- */
-/* ------------------------------------------------------------------------ */
-static uint8_t const C_cx_BrainpoolP512t1_a[] = {
-  // a:
-  // 0xAADD9DB8DBE9C48B3FD4E6AE33C9FC07CB308DB3B3C9D20ED6639CCA703308717D4D9B009BC66842AECDA12AE6A380E62881FF2F2D82C68528AA6056583A48F0
-  0xAA, 0xDD, 0x9D, 0xB8, 0xDB, 0xE9, 0xC4, 0x8B, 0x3F, 0xD4, 0xE6, 0xAE, 0x33,
-  0xC9, 0xFC, 0x07, 0xCB, 0x30, 0x8D, 0xB3, 0xB3, 0xC9, 0xD2, 0x0E, 0xD6, 0x63,
-  0x9C, 0xCA, 0x70, 0x33, 0x08, 0x71, 0x7D, 0x4D, 0x9B, 0x00, 0x9B, 0xC6, 0x68,
-  0x42, 0xAE, 0xCD, 0xA1, 0x2A, 0xE6, 0xA3, 0x80, 0xE6, 0x28, 0x81, 0xFF, 0x2F,
-  0x2D, 0x82, 0xC6, 0x85, 0x28, 0xAA, 0x60, 0x56, 0x58, 0x3A, 0x48, 0xF0
-};
-static uint8_t const C_cx_BrainpoolP512t1_b[] = {
-  // b:
-  // 0x7CBBBCF9441CFAB76E1890E46884EAE321F70C0BCB4981527897504BEC3E36A62BCDFA2304976540F6450085F2DAE145C22553B465763689180EA2571867423E
-  0x7C, 0xBB, 0xBC, 0xF9, 0x44, 0x1C, 0xFA, 0xB7, 0x6E, 0x18, 0x90, 0xE4, 0x68,
-  0x84, 0xEA, 0xE3, 0x21, 0xF7, 0x0C, 0x0B, 0xCB, 0x49, 0x81, 0x52, 0x78, 0x97,
-  0x50, 0x4B, 0xEC, 0x3E, 0x36, 0xA6, 0x2B, 0xCD, 0xFA, 0x23, 0x04, 0x97, 0x65,
-  0x40, 0xF6, 0x45, 0x00, 0x85, 0xF2, 0xDA, 0xE1, 0x45, 0xC2, 0x25, 0x53, 0xB4,
-  0x65, 0x76, 0x36, 0x89, 0x18, 0x0E, 0xA2, 0x57, 0x18, 0x67, 0x42, 0x3E
-};
-static uint8_t const C_cx_BrainpoolP512t1_p[] = {
-  // p:
-  // 0xAADD9DB8DBE9C48B3FD4E6AE33C9FC07CB308DB3B3C9D20ED6639CCA703308717D4D9B009BC66842AECDA12AE6A380E62881FF2F2D82C68528AA6056583A48F3
-  0xAA, 0xDD, 0x9D, 0xB8, 0xDB, 0xE9, 0xC4, 0x8B, 0x3F, 0xD4, 0xE6, 0xAE, 0x33,
-  0xC9, 0xFC, 0x07, 0xCB, 0x30, 0x8D, 0xB3, 0xB3, 0xC9, 0xD2, 0x0E, 0xD6, 0x63,
-  0x9C, 0xCA, 0x70, 0x33, 0x08, 0x71, 0x7D, 0x4D, 0x9B, 0x00, 0x9B, 0xC6, 0x68,
-  0x42, 0xAE, 0xCD, 0xA1, 0x2A, 0xE6, 0xA3, 0x80, 0xE6, 0x28, 0x81, 0xFF, 0x2F,
-  0x2D, 0x82, 0xC6, 0x85, 0x28, 0xAA, 0x60, 0x56, 0x58, 0x3A, 0x48, 0xF3
-};
-static uint8_t const C_cx_BrainpoolP512t1_Hp[] = {
-  0x3c, 0x4c, 0x9d, 0x05, 0xa9, 0xff, 0x64, 0x50, 0x20, 0x2e, 0x19, 0x40, 0x20,
-  0x56, 0xee, 0xcc, 0xa1, 0x6d, 0xaa, 0x5f, 0xd4, 0x2b, 0xff, 0x83, 0x19, 0x48,
-  0x6f, 0xd8, 0xd5, 0x89, 0x80, 0x57, 0xe0, 0xc1, 0x9a, 0x77, 0x83, 0x51, 0x4a,
-  0x25, 0x53, 0xb7, 0xf9, 0xbc, 0x90, 0x5a, 0xff, 0xd3, 0x79, 0x3f, 0xb1, 0x30,
-  0x27, 0x15, 0x79, 0x05, 0x49, 0xad, 0x14, 0x4a, 0x61, 0x58, 0xf2, 0x05
-};
-static uint8_t const C_cx_BrainpoolP512t1_Gx[] = {
-  // Gx:
-  // 0x640ECE5C12788717B9C1BA06CBC2A6FEBA85842458C56DDE9DB1758D39C0313D82BA51735CDB3EA499AA77A7D6943A64F7A3F25FE26F06B51BAA2696FA9035DA
-  0x64, 0x0E, 0xCE, 0x5C, 0x12, 0x78, 0x87, 0x17, 0xB9, 0xC1, 0xBA, 0x06, 0xCB,
-  0xC2, 0xA6, 0xFE, 0xBA, 0x85, 0x84, 0x24, 0x58, 0xC5, 0x6D, 0xDE, 0x9D, 0xB1,
-  0x75, 0x8D, 0x39, 0xC0, 0x31, 0x3D, 0x82, 0xBA, 0x51, 0x73, 0x5C, 0xDB, 0x3E,
-  0xA4, 0x99, 0xAA, 0x77, 0xA7, 0xD6, 0x94, 0x3A, 0x64, 0xF7, 0xA3, 0xF2, 0x5F,
-  0xE2, 0x6F, 0x06, 0xB5, 0x1B, 0xAA, 0x26, 0x96, 0xFA, 0x90, 0x35, 0xDA
-};
-static uint8_t const C_cx_BrainpoolP512t1_Gy[] = {
-  // Gy:
-  // 0x5B534BD595F5AF0FA2C892376C84ACE1BB4E3019B71634C01131159CAE03CEE9D9932184BEEF216BD71DF2DADF86A627306ECFF96DBB8BACE198B61E00F8B332
-  0x5B, 0x53, 0x4B, 0xD5, 0x95, 0xF5, 0xAF, 0x0F, 0xA2, 0xC8, 0x92, 0x37, 0x6C,
-  0x84, 0xAC, 0xE1, 0xBB, 0x4E, 0x30, 0x19, 0xB7, 0x16, 0x34, 0xC0, 0x11, 0x31,
-  0x15, 0x9C, 0xAE, 0x03, 0xCE, 0xE9, 0xD9, 0x93, 0x21, 0x84, 0xBE, 0xEF, 0x21,
-  0x6B, 0xD7, 0x1D, 0xF2, 0xDA, 0xDF, 0x86, 0xA6, 0x27, 0x30, 0x6E, 0xCF, 0xF9,
-  0x6D, 0xBB, 0x8B, 0xAC, 0xE1, 0x98, 0xB6, 0x1E, 0x00, 0xF8, 0xB3, 0x32
-};
-static uint8_t const C_cx_BrainpoolP512t1_n[] = {
-  // n:
-  // 0xAADD9DB8DBE9C48B3FD4E6AE33C9FC07CB308DB3B3C9D20ED6639CCA70330870553E5C414CA92619418661197FAC10471DB1D381085DDADDB58796829CA90069
-  0xAA, 0xDD, 0x9D, 0xB8, 0xDB, 0xE9, 0xC4, 0x8B, 0x3F, 0xD4, 0xE6, 0xAE, 0x33,
-  0xC9, 0xFC, 0x07, 0xCB, 0x30, 0x8D, 0xB3, 0xB3, 0xC9, 0xD2, 0x0E, 0xD6, 0x63,
-  0x9C, 0xCA, 0x70, 0x33, 0x08, 0x70, 0x55, 0x3E, 0x5C, 0x41, 0x4C, 0xA9, 0x26,
-  0x19, 0x41, 0x86, 0x61, 0x19, 0x7F, 0xAC, 0x10, 0x47, 0x1D, 0xB1, 0xD3, 0x81,
-  0x08, 0x5D, 0xDA, 0xDD, 0xB5, 0x87, 0x96, 0x82, 0x9C, 0xA9, 0x00, 0x69
-};
-static uint8_t const C_cx_BrainpoolP512t1_Hn[] = {
-  0xa7, 0x94, 0x58, 0x6a, 0x71, 0x84, 0x07, 0xb0, 0x95, 0xdf, 0x1b, 0x4c, 0x19,
-  0x4b, 0x2e, 0x56, 0x72, 0x3c, 0x37, 0xa2, 0x2f, 0x16, 0xbb, 0xdf, 0xd7, 0xf9,
-  0xcc, 0x26, 0x3b, 0x79, 0x0d, 0xe3, 0xa6, 0xf2, 0x30, 0xc7, 0x2f, 0x02, 0x07,
-  0xe8, 0x3e, 0xc6, 0x4b, 0xd0, 0x33, 0xb7, 0x62, 0x7f, 0x08, 0x86, 0xb7, 0x58,
-  0x95, 0x28, 0x3d, 0xdd, 0xd2, 0xa3, 0x68, 0x1e, 0xcd, 0xa8, 0x16, 0x71
-};
-
-#define C_cx_BrainpoolP512t1_h C_cofactor_1
-cx_curve_weierstrass_t const C_cx_BrainpoolP512t1 = {
-  .curve = CX_CURVE_BrainPoolP512T1,
-  .bit_size = 512,
-  .length = 64,
-
-  .a = C_cx_BrainpoolP512t1_a,
-  .b = C_cx_BrainpoolP512t1_b,
-  .p = C_cx_BrainpoolP512t1_p,
-  .Gx = C_cx_BrainpoolP512t1_Gx,
-  .Gy = C_cx_BrainpoolP512t1_Gy,
-  .n = C_cx_BrainpoolP512t1_n,
-  .h = C_cx_BrainpoolP512t1_h,
-  .Hn = C_cx_BrainpoolP512t1_Hn,
-  .Hp = C_cx_BrainpoolP512t1_Hp,
-
-};
-
-/* ========================================================================= */
-/* ========================================================================= */
-/* ===                           TWISTED EDWARDS                         === */
-/* ========================================================================= */
-/* ========================================================================= */
-
-/* ------------------------------------------------------------------------ */
-/* ---                              Ed25519                             --- */
-/* ------------------------------------------------------------------------ */
-
-static uint8_t const C_cx_Ed25519_a[] = {
-  0x7f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xec
-};
-static uint8_t const C_cx_Ed25519_d[] = {
-  // d:  0x52036cee2b6ffe738cc740797779e89800700a4d4141d8ab75eb4dca135978a3
-  0x52, 0x03, 0x6c, 0xee, 0x2b, 0x6f, 0xfe, 0x73, 0x8c, 0xc7, 0x40,
-  0x79, 0x77, 0x79, 0xe8, 0x98, 0x00, 0x70, 0x0a, 0x4d, 0x41, 0x41,
-  0xd8, 0xab, 0x75, 0xeb, 0x4d, 0xca, 0x13, 0x59, 0x78, 0xa3
-};
-static uint8_t const C_cx_Ed25519_q[] = {
-  // q:  0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffed
-  0x7f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xed
-};
-static uint8_t const C_cx_Ed25519_Hq[] = {
-  // Hq: 0x00000000000000000000000000000000000000000000000000000000000005a4
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0xa4
-};
-static uint8_t const C_cx_Ed25519_Bx[] = {
-  0x21, 0x69, 0x36, 0xd3, 0xcd, 0x6e, 0x53, 0xfe, 0xc0, 0xa4, 0xe2,
-  0x31, 0xfd, 0xd6, 0xdc, 0x5c, 0x69, 0x2c, 0xc7, 0x60, 0x95, 0x25,
-  0xa7, 0xb2, 0xc9, 0x56, 0x2d, 0x60, 0x8f, 0x25, 0xd5, 0x1a
-};
-static uint8_t const C_cx_Ed25519_By[] = {
-  0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66,
-  0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66,
-  0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x66, 0x58
-};
-static uint8_t const C_cx_Ed25519_l[] = {
-  // l:  0x1000000000000000000000000000000014DEF9DEA2F79CD65812631A5CF5D3ED
-  0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0xDE, 0xF9, 0xDE, 0xA2, 0xF7,
-  0x9C, 0xD6, 0x58, 0x12, 0x63, 0x1A, 0x5C, 0xF5, 0xD3, 0xED
-};
-static uint8_t const C_cx_Ed25519_Hl[] = {
-  // Hl: 0x0399411b7c309a3dceec73d217f5be65d00e1ba768859347a40611e3449c0f01
-  0x03, 0x99, 0x41, 0x1b, 0x7c, 0x30, 0x9a, 0x3d, 0xce, 0xec, 0x73,
-  0xd2, 0x17, 0xf5, 0xbe, 0x65, 0xd0, 0x0e, 0x1b, 0xa7, 0x68, 0x85,
-  0x93, 0x47, 0xa4, 0x06, 0x11, 0xe3, 0x44, 0x9c, 0x0f, 0x01
-};
-
-#define C_cx_Ed25519_h C_cofactor_8
-
-cx_curve_twisted_edwards_t const C_cx_Ed25519 = {
-  .curve = CX_CURVE_Ed25519,
-  .bit_size = 256,
-  .length = 32,
-
-  .a = C_cx_Ed25519_a,
-  .b = C_cx_Ed25519_d,
-  .p = C_cx_Ed25519_q,
-  .Gx = C_cx_Ed25519_Bx,
-  .Gy = C_cx_Ed25519_By,
-  .n = C_cx_Ed25519_l,
-  .h = C_cx_Ed25519_h,
-  .Hp = C_cx_Ed25519_Hq,
-  .Hn = C_cx_Ed25519_Hl,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                               Ed448                              --- */
-/* ------------------------------------------------------------------------ */
-static uint8_t const C_cx_Ed448_a[] = {
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01
-};
-static uint8_t const C_cx_Ed448_d[] = {
-  // d:
-  // 0xfffffffffffffffffffffffffffffffffffffffffffffffffffffffeffffffffffffffffffffffffffffffffffffffffffffffffffff6756
-  0x00, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xfe, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x67, 0x56
-};
-static uint8_t const C_cx_Ed448_q[] = {
-  // q:
-  // 0xfffffffffffffffffffffffffffffffffffffffffffffffffffffffeffffffffffffffffffffffffffffffffffffffffffffffffffffffff
-  0x00, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xfe, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff
-};
-static uint8_t const C_cx_Ed448_Hq[] = {
-  // Hq:
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-};
-static uint8_t const C_cx_Ed448_Bx[] = {
-  // Bx:
-  // 0x4f1970c66bed0ded221d15a622bf36da9e146570470f1767ea6de324a3d3a46412ae1af72ab66511433b80e18b00938e2626a82bc70cc05e
-  0x00, 0x4f, 0x19, 0x70, 0xc6, 0x6b, 0xed, 0x0d, 0xed, 0x22, 0x1d, 0x15,
-  0xa6, 0x22, 0xbf, 0x36, 0xda, 0x9e, 0x14, 0x65, 0x70, 0x47, 0x0f, 0x17,
-  0x67, 0xea, 0x6d, 0xe3, 0x24, 0xa3, 0xd3, 0xa4, 0x64, 0x12, 0xae, 0x1a,
-  0xf7, 0x2a, 0xb6, 0x65, 0x11, 0x43, 0x3b, 0x80, 0xe1, 0x8b, 0x00, 0x93,
-  0x8e, 0x26, 0x26, 0xa8, 0x2b, 0xc7, 0x0c, 0xc0, 0x5e
-};
-static uint8_t const C_cx_Ed448_By[] = {
-  // By:
-  // 0x693f46716eb6bc248876203756c9c7624bea73736ca3984087789c1e05a0c2d73ad3ff1ce67c39c4fdbd132c4ed7c8ad9808795bf230fa14
-  0x00, 0x69, 0x3f, 0x46, 0x71, 0x6e, 0xb6, 0xbc, 0x24, 0x88, 0x76, 0x20,
-  0x37, 0x56, 0xc9, 0xc7, 0x62, 0x4b, 0xea, 0x73, 0x73, 0x6c, 0xa3, 0x98,
-  0x40, 0x87, 0x78, 0x9c, 0x1e, 0x05, 0xa0, 0xc2, 0xd7, 0x3a, 0xd3, 0xff,
-  0x1c, 0xe6, 0x7c, 0x39, 0xc4, 0xfd, 0xbd, 0x13, 0x2c, 0x4e, 0xd7, 0xc8,
-  0xad, 0x98, 0x08, 0x79, 0x5b, 0xf2, 0x30, 0xfa, 0x14
-};
-static uint8_t const C_cx_Ed448_l[] = {
-  // l:
-  // 0x3fffffffffffffffffffffffffffffffffffffffffffffffffffffff7cca23e9c44edb49aed63690216cc2728dc58f552378c292ab5844f3
-  0x00, 0x3f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0x7c, 0xca, 0x23, 0xe9, 0xc4, 0x4e, 0xdb,
-  0x49, 0xae, 0xd6, 0x36, 0x90, 0x21, 0x6c, 0xc2, 0x72, 0x8d, 0xc5, 0x8f,
-  0x55, 0x23, 0x78, 0xc2, 0x92, 0xab, 0x58, 0x44, 0xf3
-};
-static uint8_t const C_cx_Ed448_Hl[] = {
-  // Hl:
-  0x00, 0x1a, 0x9c, 0xc1, 0x4b, 0xa3, 0xc4, 0x7c, 0x44, 0xae, 0x17, 0xcf,
-  0x72, 0xc9, 0x86, 0x0f, 0xaa, 0x10, 0x75, 0x50, 0xba, 0x94, 0xd0, 0xe8,
-  0x08, 0xf1, 0x0a, 0x98, 0xef, 0x88, 0xe1, 0x5b, 0xb4, 0x8b, 0xd8, 0x27,
-  0xde, 0x12, 0x7c, 0xe5, 0xf2, 0x3e, 0x0a, 0xa6, 0x8d, 0x30, 0xdb, 0xb9,
-  0x89, 0xbb, 0xec, 0xff, 0xc3, 0x3c, 0x3a, 0x37, 0x8c
-};
-
-#define C_cx_Ed448_h C_cofactor_4
-
-cx_curve_twisted_edwards_t const C_cx_Ed448 = {
-  .curve = CX_CURVE_Ed448,
-  .bit_size = 448,
-  .length = 57,
-
-  .a = C_cx_Ed448_a,
-  .b = C_cx_Ed448_d,
-  .p = C_cx_Ed448_q,
-  .Gx = C_cx_Ed448_Bx,
-  .Gy = C_cx_Ed448_By,
-  .n = C_cx_Ed448_l,
-  .h = C_cx_Ed448_h,
-  .Hp = C_cx_Ed448_Hq,
-  .Hn = C_cx_Ed448_Hl,
-};
-
-/* ========================================================================= */
-/* ========================================================================= */
-/* ===                             Montgomery                            === */
-/* ========================================================================= */
-/* ========================================================================= */
-
-/* ------------------------------------------------------------------------ */
-/* ---                             Curve25519                           --- */
-/* ------------------------------------------------------------------------ */
-
-static uint8_t const C_cx_Curve25519_a[] = {
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x07, 0x6d, 0x06
-};
-
-static uint8_t const C_cx_Curve25519_b[] = {
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01
-};
-static uint8_t const C_cx_Curve25519_p[] = {
-  // p:  0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffed
-  0x7f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xed
-};
-static uint8_t const C_cx_Curve25519_Hp[] = {
-  // Hp: 0x00000000000000000000000000000000000000000000000000000000000005a4
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0xa4
-};
-static uint8_t const C_cx_Curve25519_Gx[] = {
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x09
-};
-static uint8_t const C_cx_Curve25519_Gy[] = {
-  0x5f, 0x51, 0xe6, 0x5e, 0x47, 0x5f, 0x79, 0x4b, 0x1f, 0xe1, 0x22,
-  0xd3, 0x88, 0xb7, 0x2e, 0xb3, 0x6d, 0xc2, 0xb2, 0x81, 0x92, 0x83,
-  0x9e, 0x4d, 0xd6, 0x16, 0x3a, 0x5d, 0x81, 0x31, 0x2c, 0x14
-};
-static uint8_t const C_cx_Curve25519_n[] = {
-  // n:  0x1000000000000000000000000000000014DEF9DEA2F79CD65812631A5CF5D3ED
-  0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0xDE, 0xF9, 0xDE, 0xA2, 0xF7,
-  0x9C, 0xD6, 0x58, 0x12, 0x63, 0x1A, 0x5C, 0xF5, 0xD3, 0xED
-};
-static uint8_t const C_cx_Curve25519_Hn[] = {
-  // Hn: 0x0399411b7c309a3dceec73d217f5be65d00e1ba768859347a40611e3449c0f01
-  0x03, 0x99, 0x41, 0x1b, 0x7c, 0x30, 0x9a, 0x3d, 0xce, 0xec, 0x73,
-  0xd2, 0x17, 0xf5, 0xbe, 0x65, 0xd0, 0x0e, 0x1b, 0xa7, 0x68, 0x85,
-  0x93, 0x47, 0xa4, 0x06, 0x11, 0xe3, 0x44, 0x9c, 0x0f, 0x01
-};
-
-#define C_cx_Curve25519_h C_cofactor_8
-
-cx_curve_montgomery_t const C_cx_Curve25519 = {
-  .curve = CX_CURVE_Curve25519,
-  .bit_size = 256,
-  .length = 32,
-
-  .a = C_cx_Curve25519_a,
-  .b = C_cx_Curve25519_b,
-  .p = C_cx_Curve25519_p,
-  .Gx = C_cx_Curve25519_Gx,
-  .Gy = C_cx_Curve25519_Gy,
-  .n = C_cx_Curve25519_n,
-  .h = C_cx_Curve25519_h,
-  .Hp = C_cx_Curve25519_Hp,
-  .Hn = C_cx_Curve25519_Hn,
-};
-
-/* ------------------------------------------------------------------------ */
-/* ---                              Curve448                            --- */
-/* ------------------------------------------------------------------------ */
-
-static uint8_t const C_cx_Curve448_a[] = {
-  // a: 0x262a6
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x62, 0xa6
-};
-static uint8_t const C_cx_Curve448_b[] = {
-  // b:  0x01
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01
-};
-static uint8_t const C_cx_Curve448_p[] = {
-  // p:
-  // 0xfffffffffffffffffffffffffffffffffffffffffffffffffffffffeffffffffffffffffffffffffffffffffffffffffffffffffffffffff
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xfe, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff
-};
-static uint8_t const C_cx_Curve448_Hp[] = {
-  // Hp:
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-};
-static uint8_t const C_cx_Curve448_Gx[] = {
-  // Gx:
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05
-};
-static uint8_t const C_cx_Curve448_Gy[] = {
-  // Gy:
-  // 0x7d235d1295f5b1f66c98ab6e58326fcecbae5d34f55545d060f75dc28df3f6edb8027e2346430d211312c4b150677af76fd7223d457b5b1a
-  0x7d, 0x23, 0x5d, 0x12, 0x95, 0xf5, 0xb1, 0xf6, 0x6c, 0x98, 0xab, 0x6e,
-  0x58, 0x32, 0x6f, 0xce, 0xcb, 0xae, 0x5d, 0x34, 0xf5, 0x55, 0x45, 0xd0,
-  0x60, 0xf7, 0x5d, 0xc2, 0x8d, 0xf3, 0xf6, 0xed, 0xb8, 0x02, 0x7e, 0x23,
-  0x46, 0x43, 0x0d, 0x21, 0x13, 0x12, 0xc4, 0xb1, 0x50, 0x67, 0x7a, 0xf7,
-  0x6f, 0xd7, 0x22, 0x3d, 0x45, 0x7b, 0x5b, 0x1a
-};
-static uint8_t const C_cx_Curve448_n[] = {
-  // n:
-  // 0x3fffffffffffffffffffffffffffffffffffffffffffffffffffffff7cca23e9c44edb49aed63690216cc2728dc58f552378c292ab5844f3
-  0x3f, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-  0xff, 0xff, 0xff, 0xff, 0x7c, 0xca, 0x23, 0xe9, 0xc4, 0x4e, 0xdb, 0x49,
-  0xae, 0xd6, 0x36, 0x90, 0x21, 0x6c, 0xc2, 0x72, 0x8d, 0xc5, 0x8f, 0x55,
-  0x23, 0x78, 0xc2, 0x92, 0xab, 0x58, 0x44, 0xf3
-};
-static uint8_t const C_cx_Curve448_Hn[] = {
-  // Hn:
-  0x1a, 0x9c, 0xc1, 0x4b, 0xa3, 0xc4, 0x7c, 0x44, 0xae, 0x17, 0xcf, 0x72,
-  0xc9, 0x86, 0x0f, 0xaa, 0x10, 0x75, 0x50, 0xba, 0x94, 0xd0, 0xe8, 0x08,
-  0xf1, 0x0a, 0x98, 0xef, 0x88, 0xe1, 0x5b, 0xb4, 0x8b, 0xd8, 0x27, 0xde,
-  0x12, 0x7c, 0xe5, 0xf2, 0x3e, 0x0a, 0xa6, 0x8d, 0x30, 0xdb, 0xb9, 0x89,
-  0xbb, 0xec, 0xff, 0xc3, 0x3c, 0x3a, 0x37, 0x8c
-};
-
-#define C_cx_Curve448_h C_cofactor_4
-
-cx_curve_montgomery_t const C_cx_Curve448 = {
-  .curve = CX_CURVE_Curve448,
-  .bit_size = 448,
-  .length = 56,
-
-  .a = C_cx_Curve448_a,
-  .b = C_cx_Curve448_b,
-  .p = C_cx_Curve448_p,
-  .Gx = C_cx_Curve448_Gx,
-  .Gy = C_cx_Curve448_Gy,
-  .n = C_cx_Curve448_n,
-  .h = C_cx_Curve448_h,
-  .Hp = C_cx_Curve448_Hp,
-  .Hn = C_cx_Curve448_Hn,
-};
-
-/* ========================================================================= */
-/* ========================================================================= */
-/* ===                                ALL                                === */
-/* ========================================================================= */
-/* ========================================================================= */
-
-cx_curve_domain_t const *const C_cx_allCurves[] = {
-
-  // secp
-  (const cx_curve_domain_t *)&C_cx_secp256k1,
-  (const cx_curve_domain_t *)&C_cx_secp256r1,
-  (const cx_curve_domain_t *)&C_cx_secp384r1,
-  (const cx_curve_domain_t *)&C_cx_secp521r1,
-
-  // Brainpool/NIST
-  (const cx_curve_domain_t *)&C_cx_BrainpoolP256r1,
-  (const cx_curve_domain_t *)&C_cx_BrainpoolP256t1,
-  (const cx_curve_domain_t *)&C_cx_BrainpoolP320r1,
-  (const cx_curve_domain_t *)&C_cx_BrainpoolP320t1,
-  (const cx_curve_domain_t *)&C_cx_BrainpoolP384r1,
-  (const cx_curve_domain_t *)&C_cx_BrainpoolP384t1,
-  (const cx_curve_domain_t *)&C_cx_BrainpoolP512r1,
-  (const cx_curve_domain_t *)&C_cx_BrainpoolP512t1,
-
-  // TwEd
-  (const cx_curve_domain_t *)&C_cx_Ed25519,
-  (const cx_curve_domain_t *)&C_cx_Ed448,
-
-  (const cx_curve_domain_t *)&C_cx_Curve25519,
-  (const cx_curve_domain_t *)&C_cx_Curve448
-};
-
-//-----------------------------------------------------------------------------
-
-// Almost duplicated from cx_ec.c one because some curves id have changed!
-int cx_nid_from_curve(cx_curve_t curve)
-{
-  int nid;
-
-  switch (curve) {
-  case CX_CURVE_SECP256K1:
-    nid = NID_secp256k1;
-    break;
-  case CX_CURVE_SECP256R1:
-    nid = NID_X9_62_prime256v1;
-    break;
-  case CX_CURVE_SECP384R1:
-    nid = NID_secp384r1;
-    break;
-  case CX_CURVE_Ed25519:
-    nid = NID_ED25519;
-    break;
-  case CX_CURVE_BrainPoolP256R1:
-    nid = NID_brainpoolP256r1;
-    break;
-  case CX_CURVE_BrainPoolP256T1:
-    nid = NID_brainpoolP256t1;
-    break;
-  case CX_CURVE_BrainPoolP320R1:
-    nid = NID_brainpoolP320r1;
-    break;
-  case CX_CURVE_BrainPoolP320T1:
-    nid = NID_brainpoolP320t1;
-    break;
-  case CX_CURVE_BrainPoolP384R1:
-    nid = NID_brainpoolP384r1;
-    break;
-  case CX_CURVE_BrainPoolP384T1:
-    nid = NID_brainpoolP384t1;
-    break;
-  case CX_CURVE_BrainPoolP512R1:
-    nid = NID_brainpoolP512r1;
-    break;
-  case CX_CURVE_BrainPoolP512T1:
-    nid = NID_brainpoolP512t1;
-    break;
-#if 0
-  case CX_CURVE_SECP521R1:
-    nid = NID_secp521r1;
-    break;
-#endif
-  default:
-    nid = -1;
-    errx(1, "cx_nid_from_curve unsupported curve");
-    break;
-  }
-  return nid;
-}
-
-const cx_curve_domain_t *cx_ecdomain(cx_curve_t curve)
-{
-  size_t i;
-  for (i = 0; i < sizeof(C_cx_allCurves) / sizeof(cx_curve_domain_t *); i++) {
-    if (C_cx_allCurves[i]->curve == curve) {
-      return C_cx_allCurves[i];
-    }
-  }
-  return NULL;
-}
-
-cx_err_t sys_cx_ecdomain_parameters_length(cx_curve_t curve, size_t *length)
-{
-  const cx_curve_domain_t *domain;
-
-  domain = cx_ecdomain(curve);
-  if (domain == NULL) {
-    return CX_EC_INVALID_CURVE;
-  }
-
-  *length = domain->length;
-  return CX_OK;
-}
-
-cx_err_t sys_cx_ecdomain_size(cx_curve_t curve, size_t *length)
-{
-  const cx_curve_domain_t *domain;
-
-  domain = cx_ecdomain(curve);
-  if (domain == NULL) {
-    return CX_EC_INVALID_CURVE;
-  }
-
-  *length = domain->bit_size;
-  return CX_OK;
-}
-
-static void const *sys_cx_ecdomain_param_ptr(const cx_curve_domain_t *domain,
-                                             cx_curve_dom_param_t id)
-{
-  switch (id) {
-  case CX_CURVE_PARAM_A:
-    return domain->a;
-  case CX_CURVE_PARAM_B:
-    return domain->b;
-  case CX_CURVE_PARAM_Field:
-    return domain->p;
-  case CX_CURVE_PARAM_Gx:
-    return domain->Gx;
-  case CX_CURVE_PARAM_Gy:
-    return domain->Gy;
-  case CX_CURVE_PARAM_Order:
-    return domain->n;
-  case CX_CURVE_PARAM_Cofactor:
-    return domain->h;
-  default:
-    break;
-  }
-  return NULL;
-}
-
-cx_err_t sys_cx_ecdomain_parameter(cx_curve_t curve, cx_curve_dom_param_t id,
-                                   uint8_t *p, uint32_t p_len)
-{
-  void const *ptr;
-  const cx_curve_domain_t *domain;
-
-  domain = cx_ecdomain(curve);
-  if (domain == NULL) {
-    return CX_EC_INVALID_CURVE;
-  }
-
-  ptr = sys_cx_ecdomain_param_ptr(domain, id);
-  if ((ptr == NULL) || (domain->length > p_len)) {
-    return CX_INVALID_PARAMETER;
-  }
-
-  if (id == CX_CURVE_PARAM_Cofactor) {
-    memmove(p, ptr, 4);
-  } else {
-    memmove(p, ptr, domain->length);
-  }
-  return CX_OK;
-}
-
-cx_err_t sys_cx_ecdomain_parameter_bn(cx_curve_t curve, cx_curve_dom_param_t id,
-                                      cx_bn_t bn_p)
-{
-  cx_err_t error;
-  cx_mpi_t *p;
-  void const *ptr;
-  const cx_curve_domain_t *domain;
-
-  domain = cx_ecdomain(curve);
-  if (domain == NULL) {
-    return CX_EC_INVALID_CURVE;
-  }
-  CX_CHECK(cx_bn_to_mpi(bn_p, &p));
-
-  ptr = sys_cx_ecdomain_param_ptr(domain, id);
-  if (ptr == NULL) {
-    return CX_INVALID_PARAMETER;
-  }
-  // Copy the content of ptr into BN:
-  if (id == CX_CURVE_PARAM_Cofactor) {
-    BN_bin2bn(ptr, 4, p);
-  } else {
-    BN_bin2bn(ptr, domain->length, p);
-  }
-end:
-  return error;
-}
-
-cx_err_t sys_cx_ecdomain_generator(cx_curve_t curve, uint8_t *Gx, uint8_t *Gy,
-                                   size_t len)
-{
-  const cx_curve_domain_t *domain;
-
-  domain = cx_ecdomain(curve);
-  if (domain == NULL) {
-    return CX_EC_INVALID_CURVE;
-  }
-  if (domain->length > len) {
-    return CX_INVALID_PARAMETER;
-  }
-  memmove(Gx, domain->Gx, domain->length);
-  memmove(Gy, domain->Gy, domain->length);
-
-  return CX_OK;
-}
-
-cx_err_t sys_cx_ecdomain_generator_bn(cx_curve_t curve, cx_ecpoint_t *ec_P)
-{
-  cx_err_t error;
-  cx_mpi_ecpoint_t P;
-  const cx_curve_domain_t *domain;
-
-  CX_CHECK(cx_bn_locked());
-
-  domain = cx_ecdomain(curve);
-  if (domain == NULL) {
-    return CX_EC_INVALID_CURVE;
-  }
-  if ((curve != ec_P->curve) ||
-      (cx_mpi_ecpoint_from_ecpoint(&P, ec_P) != CX_OK)) {
-    return CX_EC_INVALID_POINT;
-  }
-  BN_bin2bn(domain->Gx, domain->length, P.x);
-  BN_bin2bn(domain->Gy, domain->length, P.y);
-  cx_mpi_set_u32(P.z, 1);
-end:
-  return error;
-}
