+++ src/bolos/cx_ed25519.c
@@ -8,6 +8,7 @@
 
 #include "cx_ec.h"
 #include "cx_ed25519.h"
+#include "cx_errors.h"
 #include "cx_utils.h"
 
 static const char *constant_q = "5789604461865809771178549250434395392663499233"
@@ -172,6 +173,31 @@ int sys_cx_edward_compress_point(cx_curve_t curve, uint8_t *P, size_t P_len)
   P[0] = 0x02;
 
   return 0;
+}
+
+int sys_cx_edward_compress_point_no_throw(cx_curve_t curve, uint8_t *P,
+                                          size_t P_len)
+{
+  const cx_curve_twisted_edward_t *domain;
+  uint8_t *x, *y;
+  size_t size;
+
+  domain = (const cx_curve_twisted_edward_t *)cx_ecfp_get_domain(curve);
+  size = domain->length;
+  if (curve != CX_CURVE_Ed25519 || P_len < (1 + 2 * size)) {
+    // errx(1, "cx_edward_compress_point: invalid parameters");
+    return CX_INVALID_PARAMETER;
+  }
+
+  x = P + 1;
+  y = P + 1 + size;
+
+  cx_compress(x, y, size);
+  memcpy(x, y, size);
+
+  P[0] = 0x02;
+
+  return CX_OK;
 }
 
 static int xrecover(BIGNUM *x, BIGNUM *y)
