+++ src/bolos/cx_eddsa.c
@@ -30,20 +30,20 @@ int ED25519_verify(const uint8_t *message, size_t message_len,
 void ED25519_public_from_private(uint8_t out_public_key[32],
                                  const uint8_t private_key[32]);
 
-int sys_cx_eddsa_sign(const cx_ecfp_private_key_t *pvkey,
-                      int mode __attribute__((unused)), cx_md_t hashID,
-                      const unsigned char *hash, unsigned int hash_len,
-                      const unsigned char *ctx __attribute__((unused)),
-                      unsigned int ctx_len __attribute__((unused)),
-                      unsigned char *sig, unsigned int sig_len,
-                      unsigned int *info __attribute__((unused)))
+int sys_cx_eddsa_sign_no_throw(const cx_ecfp_private_key_t *pvkey,
+                               int mode __attribute__((unused)), cx_md_t hashID,
+                               const unsigned char *hash, unsigned int hash_len,
+                               const unsigned char *ctx __attribute__((unused)),
+                               unsigned int ctx_len __attribute__((unused)),
+                               unsigned char *sig, unsigned int sig_len,
+                               unsigned int *info __attribute__((unused)))
 {
   uint8_t public_key[32];
 
-  /* Only SHA-512 is supported in Speculos, as it is the only supported hash
+  /* Only SHA-512 is supported in Speculos, as it was the only supported hash
    * in OpenSSL */
   if (hashID != CX_SHA512) {
-    return -1;
+    return CX_INVALID_PARAMETER;
   }
   if (sig_len < 64) {
     return CX_KO;
@@ -54,6 +54,35 @@ int sys_cx_eddsa_sign(const cx_ecfp_private_key_t *pvkey,
   }
   ED25519_public_from_private(public_key, pvkey->d);
   if (ED25519_sign(sig, hash, hash_len, public_key, pvkey->d) == 1) {
+    return CX_OK;
+  }
+  return CX_KO;
+}
+
+int sys_cx_eddsa_sign(const cx_ecfp_private_key_t *pvkey,
+                      int mode __attribute__((unused)), cx_md_t hashID,
+                      const unsigned char *hash, unsigned int hash_len,
+                      const unsigned char *ctx __attribute__((unused)),
+                      unsigned int ctx_len __attribute__((unused)),
+                      unsigned char *sig, unsigned int sig_len,
+                      unsigned int *info __attribute__((unused)))
+{
+  uint8_t public_key[32];
+
+  /* Only SHA-512 is supported in Speculos, as it was the only supported hash
+   * in OpenSSL */
+  if (hashID != CX_SHA512) {
+    return CX_INVALID_PARAMETER;
+  }
+  if (sig_len < 64) {
+    return CX_KO;
+  }
+  /* Key must be a Ed25519 private key */
+  if (pvkey->curve != CX_CURVE_Ed25519 || pvkey->d_len != 32) {
+    return CX_KO;
+  }
+  ED25519_public_from_private(public_key, pvkey->d);
+  if (ED25519_sign(sig, hash, hash_len, public_key, pvkey->d) == OPEN_SSL_OK) {
     return _EDD_SIG_T8;
   }
   return CX_KO;
@@ -83,7 +112,7 @@ int sys_cx_eddsa_verify(const cx_ecfp_public_key_t *pu_key,
   /* pass a compressed key to ED25519_verify */
   memcpy(&pub, pu_key, sizeof(pub));
   if (pub.W[0] == 0x04) {
-    if (sys_cx_edward_compress_point(pub.curve, pub.W, pub.W_len) != 0) {
+    if (sys_cx_edward_compress_point(pub.curve, pub.W, pub.W_len) != CX_OK) {
       return 0;
     }
   }
