+++ src/bolos/cx_wrap_ossl.h
@@ -6,14 +6,15 @@
 #define NID_BLS12_381_G1 666
 
 /* UGLY return codes of openSSL, when OK=1 and KO=0 :( */
-#define OPEN_SSL_KO 0
-#define OPEN_SSL_OK 1
-
+#define OPEN_SSL_KO               0
+#define OPEN_SSL_OK               1
+#define OPEN_SSL_UNEXPECTED_ERROR 0xcaca055L
 /* openssl wrappers*/
 int cx_generic_curve(const cx_curve_weierstrass_t *i_weier, BN_CTX *ctx,
                      EC_GROUP **o_group);
 
 int nid_from_curve(cx_curve_t curve);
+#define cx_nid_from_curve(a) nid_from_curve(a)
 
 int spec_cx_ecfp_encode_sig_der(unsigned char *sig, unsigned int sig_len,
                                 unsigned char *r, unsigned int r_len,
