+++ src/bolos/cxlib.h
@@ -156,9 +156,11 @@ cx_err_t sys_cx_bn_mod_pow2(cx_bn_t bn_r, const cx_bn_t bn_a, const uint8_t *e,
 cx_err_t sys_cx_bn_is_prime(const cx_bn_t bn_x, bool *prime);
 cx_err_t sys_cx_bn_next_prime(const cx_bn_t bn_x);
 
-// cx_ecdomain.c
 int cx_nid_from_curve(cx_curve_t curve);
-const cx_curve_domain_t *cx_ecdomain(cx_curve_t curve);
+
+/* TODO: remove those calls to cx_ec and replace it*/
+#define cx_ecdomain(curve) cx_ecfp_get_domain(curve)
+
 cx_err_t sys_cx_ecdomain_parameters_length(cx_curve_t curve, size_t *length);
 cx_err_t sys_cx_ecdomain_size(cx_curve_t curve, size_t *length);
 cx_err_t sys_cx_ecdomain_parameter(cx_curve_t curve, cx_curve_dom_param_t id,
